==> prep/02_feed/1a_GAEZ_production.Rmd <==
---
title: "Calculate total tonnes of production in GAEZ data"
output: html_document
editor_options: 
  chunk_output_type: console
---

# Summary

This should be run if there are changes to the GAEZ production data.

Here we calculate total crop production using the GAEZ data adjusted to mean of 2019-2021 production. These data include all growing systems (intensive, irrigated, etc.) and are in units of metric tonnes.

This is used in subsequent script to estimate the proportion of each crop going to each animal system (for this paper, only Atlantic salmon) feed in each country.

# Data sources

**Reference**: Grogan, D., Frolking, S., Wisser, D. et al. Global gridded crop harvested area, production, yield, and monthly physical area data circa 2015. Sci Data 9, 15 (2022). https://doi.org/10.1038/s41597-021-01115-2

**Downloaded**: 07/14/2023

**Description**: "Here we provide an update to global gridded annual and monthly crop datasets. This new dataset uses the crop categories established by the Global Agro-Ecological Zones (GAEZ) Version 3 model, which is based on the Food and Agricultural Organization of the United Nations (FAO) crop production data. We used publicly available data from the FAOSTAT database as well as GAEZ Version 4 global gridded dataset to generate circa 2015 annual crop harvested area, production, and yields by crop production system (irrigated and rainfed) for 26 crops and crop categories globally at 5-minute resolution. We additionally used available data on crop rotations, cropping intensity, and planting and harvest dates to generate monthly gridded cropland data for physical areas for the 26 crops by production system. These data are in standard georeferenced gridded format, and can be used by any global hydrology, land surface, or other earth system model that requires gridded annual or monthly crop data inputs."

**Download Link**: https://gaez.fao.org/pages/data-access-download


**Reference**: FAOSTAT (2023) ‘Crops and livestock products’. Available at: https://www.fao.org/faostat/en/#data/QCL (Accessed: 14 March 2023).

**Downloaded**: 03/14/2023

**Description**: FAOSTAT crop production statistics 

# Methods

## Determine total crop grown in each region

```{r}
feed_folder <- file.path(here("prep/02_feed"))

library(here)
library(tidyverse)
library(countrycode)
library(strex)
library(terra)

source("src/directories.R")

```

## Extract by region the crop production from the raster maps.

```{r}

regions <- read_csv(here("data/spatial/output/food_rgns.csv")) # these are regions originally derived from GADM36 datasets, but I poached them from https://knb.ecoinformatics.org/view/doi%3A10.5063%2FF1FN14NK
 
crop_tonnes <- list.files(file.path(gaez_dir, "/scaled_prod_maps_2021"), pattern = "_Total_scaled.tif", full=TRUE)

crop_stack <- terra::rast(crop_tonnes)

names(crop_stack) <- str_after_last(sources(crop_stack), "/")

zones <- terra::rast(file.path(rdsi_raw_data_dir, "food-systems-project/land_eez_rgns.tif")) # taken from https://knb.ecoinformatics.org/view/doi%3A10.5063%2FF1FN14NK

crop_production <- terra::zonal(crop_stack, zones, fun="sum", na.rm=TRUE, progress="text")

crop_production_df <- data.frame(crop_production) %>%
  gather("GAEZ_category", "tonnes_producing_crop", -1) %>%
  rename("ID_0" = land_eez_rgns) %>%
  left_join(regions, by="ID_0") %>%
  mutate(GAEZ_category = gsub("crop_", "", GAEZ_category)) %>%
  mutate(GAEZ_category = gsub("_Total_scaled.tif", "", GAEZ_category)) %>%
  dplyr::select(iso3c_producing=iso3c, GAEZ_category, tonnes_producing_crop) %>%
  mutate(tonnes_producing_crop = ifelse(is.na(tonnes_producing_crop), 0, tonnes_producing_crop)) # replace NAs with 0

```

## Check and save: should be about equal to 2021 FAO production

```{r}

check <- crop_production_df %>%
  group_by(GAEZ_category) %>% 
  summarize(our_tonnes = sum(tonnes_producing_crop, na.rm=TRUE)) %>%
  data.frame()
  
fao_translate <- vroom::vroom(here("prep/01_crops_prep/data/GAEZ_names.csv")) %>%
  dplyr::select(GAEZ_category, item_code) %>%
  unique()

fao <- read_csv(here("prep/01_crops_prep/data/FAOSTAT_crop_production_2015_2021.csv")) %>% 
  filter(year %in% c(2019, 2020, 2021)) %>%
  filter(area != "China",
         area_code < 5000) %>%
 # dplyr::select(area, item_code, value, year) %>%
  left_join(fao_translate, by="item_code") %>% 
  group_by(GAEZ_category, year) %>%
  summarise(total = sum(value, na.rm = TRUE)) %>%
  ungroup() %>%
  group_by(GAEZ_category) %>%
  summarize(total = mean(total, na.rm=TRUE)) %>%
  ungroup() %>%
  filter(!is.na(GAEZ_category))
  
compare <- check %>% left_join(fao, by="GAEZ_category") %>%
  mutate(prop = our_tonnes/total)
compare
sum(compare$our_tonnes, na.rm=TRUE)/sum(compare$total, na.rm = TRUE) # oilpalmfruit, millet, and stimulants are a little problematic (stimulants is the worst ~50%...). However, we don't use the truly horrible ones for this feed analysis.

write_csv(crop_production_df, here("prep/02_feed/data/GAEZcrop_production.csv"))

test <- read.csv(here("prep/02_feed/data/GAEZcrop_production.csv"))

```
