==> prep/02_feed/2_aquaculture_diet_composition.Rmd <==
---
title: "Aquaculture diet composition wrangling"
output: html_document
editor_options: 
  chunk_output_type: console
---

# Summary 
Wrangling and aggregating aquaculture diet composition data.

```{r}
library(here)
source(here("src/fxns.R"))

library(tidyverse)
library(readxl)

regions <- read_csv(here("data/spatial/output/food_rgns.csv")) %>%
  select(iso3c) %>%
  unique()

```

# Data Source 

**Reference**: Aas, T.S., Åsgård, T. and Ytrestøyl, T. (2022) ‘Utilization of feed resources in the production of Atlantic salmon (Salmo salar) in Norway: An update for 2020’, Aquaculture Reports, 26, p. 101316. Available at: https://doi.org/10.1016/j.aqrep.2022.101316.


# Methods

Read in diet composition data, derived from Aas et al., 2022

```{r}

salmon_plant <- read_xlsx(here("prep/02_feed/data/diet_composition/salmon_aquaculture_diet_composition.xlsx"), sheet = "2020") %>%
  mutate(raw_name = ifelse(raw_name == "camelina oil", "rapeseed and camelina oil", raw_name),
         raw_name = ifelse(raw_name == "rapeseed oil", "rapeseed and camelina oil", raw_name)) %>%
  group_by(species, production_system, raw_name_edit) %>%
  summarise(per_diet = sum(per_diet, na.rm = TRUE)) %>%
  ungroup() %>%
  mutate(prop_diet = per_diet * 0.01) %>%
  merge(regions) %>%
  select(iso3c, raw_name = raw_name_edit, prop_diet, production_system, species) %>%
  mutate(diet = "plant-dominant")

salmon_fish <-  read_xlsx(here("prep/02_feed/data/diet_composition/salmon_aquaculture_diet_composition.xlsx"), sheet = "fish-dominant") %>%
  mutate(raw_name = ifelse(raw_name == "camelina oil", "rapeseed and camelina oil", raw_name),
         raw_name = ifelse(raw_name == "rapeseed oil", "rapeseed and camelina oil", raw_name)) %>%
  group_by(species, production_system, raw_name_edit) %>%
  summarise(per_diet = sum(per_diet, na.rm = TRUE)) %>%
  ungroup() %>%
  mutate(prop_diet = per_diet * 0.01) %>%
  merge(regions) %>%
  select(iso3c, raw_name = raw_name_edit, prop_diet, production_system, species) %>%
  mutate(diet = "fish-dominant")

salmon <- rbind(salmon_plant, salmon_fish)



write.csv(salmon, here("prep/02_feed/data/aquaculture_diet_composition.csv"), row.names = FALSE)


```
