==> prep/02_feed/3_aquaculture_diet_total_consumption.Rmd <==
---
title: "Calculating aquaculture diet consumption in tonnes"
author: "Gage Clawson"
date: "03/20/2023"
output: html_document
editor_options: 
  chunk_output_type: console
---

## Summary 

Use following data to calculate tonnes of each feedstuff item consumed by each aquaculture system in each country:

* Tonnes of production: From FAOSTAT
* Diet composition: From Aas et al. 2022 and Cottrell et al. 2024 in review
* feed conversion ratio: From Tacon & Metian 2008; 2015


## Data Sources

**Reference**: FAO (2023) ‘Fishery and Aquaculture Statistics. Global Fish Trade - All partners aggregated 1976-2021’. FishStatJ.

**Downloaded**: 11/03/2022

**Reference**: FAOSTAT (2023) ‘Crops and livestock products’. Available at: https://www.fao.org/faostat/en/#data/QCL (Accessed: 14 March 2023).

**Downloaded**: 03/14/2023

**Reference**: Aas, T.S., Åsgård, T. and Ytrestøyl, T. (2022) ‘Utilization of feed resources in the production of Atlantic salmon (Salmo salar) in Norway: An update for 2020’, Aquaculture Reports, 26, p. 101316. Available at: https://doi.org/10.1016/j.aqrep.2022.101316.

**Reference**: Tacon, A.G.J. and Metian, M. (2015) ‘Feed Matters: Satisfying the Feed Demand of Aquaculture’, Reviews in Fisheries Science & Aquaculture, 23(1), pp. 1–10. Available at: https://doi.org/10.1080/23308249.2014.987209.

**Reference**: Tacon, A.G.J. and Metian, M. (2008) ‘Global overview on the use of fish meal and fish oil in industrially compounded aquafeeds: Trends and future prospects’, Aquaculture, 285(1–4), pp. 146–158. Available at: https://doi.org/10.1016/j.aquaculture.2008.08.015.



```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)


library(here)
library(stringi)
library(countrycode)

scenario_year <- 2020

source(here("src/fxns.R"))
source(here("src/directories.R"))
```


```{r}

production <- readRDS(here("data/tidy_data/production-data/aquaculture_production_tidy.rds")) 


salmon_production <- production %>%
  filter(species == "Atlantic salmon" & year == scenario_year,
         value > 0) %>%
  mutate(species = "salmon")


diet <- read_csv(here("prep/02_feed/data/aquaculture_diet_composition.csv"))

fcrs <- read.csv(here("prep/02_feed/data/FCR.csv")) %>%
  mutate(species = "salmon") %>% # should i just use 1.2 here like Rich did? 
    mutate(country = trimws(country))


```

Import the feed conversion ratios and join to production. 

Feed conversion ratios were sourced from Tacon and Metian 2008 which has country specific information for salmon feed conversion ratios.

Where country information from tacon and metian is not available use global average for that country. And get absolute feed demand. To do this we use the min eFCR given the Tacon and Metian data is old and there is a general upward trend in feed efficiency through time. However, for countries where no specific fcr data is provided, give the mean Global FCR.


```{r}
## add in FCR scenarios here

feed_demand <- 
  salmon_production %>%
  mutate(country = trimws(country)) %>%
  left_join(fcrs, by = c("country", "species")) %>%
  mutate(eFCR_min = if_else(is.na(eFCR_min), true = fcrs %>% filter(country=="Globe") %>% pull(eFCR_mean), false = eFCR_min),
         eFCR_max = if_else(is.na(eFCR_max), true = fcrs %>% filter(country=="Globe") %>% pull(eFCR_mean), false = eFCR_max),
         eFCR_mean = if_else(is.na(eFCR_mean), true = fcrs %>% filter(country=="Globe") %>% pull(eFCR_mean), false = eFCR_mean)) %>% 
  mutate(feed_demand = value*eFCR_min) %>%
  mutate(feed_demand_efficient = value*(eFCR_min*(1-0.1)))

saveRDS(feed_demand, here("prep/02_feed/data/demand/global_feed_demand.rds"))
feed_demand <- readRDS(here("prep/02_feed/data/demand/global_feed_demand.rds"))
```


Now calculate ingredient demand per diet and for each salmon producing country. Global demand is affected by the among country differences in FCR.

```{r}

tonnes_consumption <- feed_demand %>%
  left_join(diet, by=c("iso3c", "species")) %>%
  mutate(ingredient_demand = feed_demand * prop_diet) %>%
  mutate(ingredient_demand_efficient = feed_demand_efficient * prop_diet) %>%
  mutate(product = "meat") %>%
  dplyr::select(iso3c, animal= species, system = production_system, product, ingredient = raw_name, ingredient_demand, ingredient_demand_efficient, diet) 

write_csv(tonnes_consumption, here("prep/02_feed/data/demand/total_aquaculture_ingredient_demand.csv")) 

tonnes_consumption %>%
  group_by(ingredient, diet) %>%
  summarise(total_ingredient_demand = sum(ingredient_demand, na.rm = TRUE),
            total_ingredient_demand_efficient = sum(ingredient_demand_efficient, na.rm = TRUE)) %>%
  ungroup() %>%
  write.csv(here("prep/02_feed/data/demand/total_global_ingredient_demand.csv"), row.names = FALSE)

test <- read.csv(here("prep/02_feed/data/demand/total_global_ingredient_demand.csv"))
```
