==> prep/02_feed/4a_crop_ingredient_allocation.Rmd <==
---
title: "Account for crop trade, allocation factors, and rasterize!"
output: html_document
editor_options: 
  chunk_output_type: console
---

# Summary 

GOAL: Get proportion of each GAEZ crop that is consumed by salmon aquaculture in each country. Allocate to raw weight materials using mass, economic, and energetic allocation rates.

Outputs: a raster for each crop describing the amount of each crop produced (tonnes) and area based disturbance (km2 equivalents) used by salmon aquaculture systems for each scenario.

```{r}

library(here)
library(tidyverse)
library(countrycode)
library(tidytext)
library(knitr)
library(strex)
library(readxl)
library(terra)
library(tidyterra) 

source(here("src/fxns.R"))
source(here("src/directories.R"))

```

## Data Sources:

 - GAEZ, FAO crop production, and FAO trade data from scripts before
 
**Reference**: Cottrell, R.S. et al. (2024) ‘Raw material origins matter as much as feed composition for the environmental footprint of aquaculture feeds’, Manuscript in review [Preprint].

**Description**: We use the raw material allocation methods created from this paper. 

# Methods

## STEP 1: determine amount of crop produced in each country for each food system.

Trade proportion data:

```{r}

trade_data <- read_csv(here("prep/02_feed/data/FAO_GAEZcrop_trade_data_10-02-2024.csv")) 

```


Total demand by animals per country in GAEZ crops

```{r}

consumption <- read_csv(here("prep/02_feed/data/demand/total_aquaculture_ingredient_demand.csv")) %>%
  mutate(animal_system = paste(animal, system, product, sep="_")) %>%
  rename(source_ingredient = ingredient) %>%
  pivot_longer(cols = c("ingredient_demand", "ingredient_demand_efficient"), names_to = "fcr_type", values_to = "consuming_ingredient_consumed_tonnes") %>%
  mutate(fcr_type = ifelse(fcr_type == "ingredient_demand", "regular", "efficient")) %>% 
  select(iso3c_consuming = iso3c, animal_system, diet, source_ingredient, fcr_type, consuming_ingredient_consumed_tonnes) 

## need to match `consumption` ingredients to equivalent crop.. need look up table 

feed_cats <- read_xlsx(here("data/tidy_data/plant_ingredient_codes.xlsx")) %>% ## this table was created manually. It matches our ingredient names from Aas et al. 2022 to GAEZ raw material crop names. Taken from our IMAS Sustainable Aquafeed Project - Mapping directory on OneDrive, specifically the "allocation_new" folder. 
  dplyr::select(-reference)
  


consumption_spam_gaez <- consumption %>%
  left_join(feed_cats, by = c("source_ingredient" = "ingredient")) %>%
  filter(!is.na(raw_material)) # filter out fish and microingredients, handle these later

setdiff(unique(consumption$source_ingredient), unique(feed_cats$ingredient)) ## this is only fish ingredients and microingredients/other, which we don't care about for crops (because not crops/not including)


setdiff(unique(feed_cats$ingredient), unique(consumption$source_ingredient)) # [1]  corn starch, pea starch, and lupin kernel meal... not included for biodiversity project, so ok

```


Join consumption by country and crop to trade data to get likely country of origin regardless of country of consumption.

```{r}
# check that all feed products match up: should be nothing showing up here.. if it is then add it to lookup table
setdiff(consumption_spam_gaez$gaez_code, trade_data$GAEZ_category) # character(0)


ingredient_production_location <- left_join(trade_data, consumption_spam_gaez, by=c("iso3c_consuming", "GAEZ_category" = "gaez_code")) %>%
  filter(!is.na(animal_system)) %>%  #crops not fed to animal system in given country
  mutate(consuming_ingredient_consumed_tonnes = ifelse(is.na(consuming_ingredient_consumed_tonnes), 0, consuming_ingredient_consumed_tonnes)) %>%
  mutate(producing_ingredient_consumed_tonnes = consuming_ingredient_consumed_tonnes * prop_of_supply)


write.csv(ingredient_production_location, here("prep/02_feed/data/demand/ingredient_demand_production_locations_10-02-2024.csv"), row.names = FALSE)
```


Take a look at the data 

```{r}

feed_source <- ingredient_production_location %>%
  group_by(iso3c_producing, diet, fcr_type, animal_system, GAEZ_category, source_ingredient) %>%
  summarize(consumed_ingredient_tonnes = sum(producing_ingredient_consumed_tonnes)) %>% 
  ungroup()

tmp <- filter(feed_source, animal_system == "salmon_aquaculture_meat") %>%
    group_by(source_ingredient, diet, fcr_type) %>%
  slice_max(order_by = consumed_ingredient_tonnes, n = 15) %>%
  ungroup() %>%
  filter(consumed_ingredient_tonnes > 0)


ggplot(data = tmp %>%
         filter(diet == "fish-dominant"), aes(x = consumed_ingredient_tonnes, y = reorder_within(iso3c_producing, consumed_ingredient_tonnes, GAEZ_category))) + 
  geom_bar(stat = "identity") + 
  facet_grid(fcr_type~source_ingredient, scales = "free") + 
  labs(x = "Tonnes of ingredient produced", y = "Country") + 
  theme_bw() + 
  theme(axis.text.x = element_text(angle = 30, vjust = 0.5)) 


```
 

## STEP 2: Use allocation methods to account for crop loss and coproducts to get a better idea of the amount of crops actually going to aquaculture feeds (instead of ingredients)

Ok now we need to convert the ingredient to crop raw material using the conversion factors that Rich has put together.. we do mass, economic, and energetic allocation  


```{r}
## join ingredient demand and allocation data to generate crop raw material demand
 
ingredient_production <- ingredient_production_location %>%
  group_by(iso3c_producing, GAEZ_category, diet, fcr_type, source_ingredient, raw_material, FAOSTAT_name, FAO_code, map_spam_code) %>%
  summarise(producing_ingredient_consumed_tonnes = sum(producing_ingredient_consumed_tonnes, na.rm = TRUE)) %>%
  ungroup()
 
allocation_factors <- read_xlsx(here("data/tidy_data/allocation/crop_ingredient_allocation_factors.xlsx")) %>% 
  dplyr::select(item, ingredient, FAO_code, map_spam_code, gaez_code, ge_allocation_factor, mass_allocation_factor, econ_allocation_factor) %>%
  distinct()


# join the ingredient demand by diet to the producer data so for each ingredient there will be a raw material demand per source, fcr type, and diet scenario. Create new raw_material demand column by multiplying total ingredient demand by the conversion factor (i.e. the ingredient extraction rate from the raw material in the first place)

crop_raw_material_demand <- ingredient_production_location %>%
  left_join(allocation_factors, by = c("source_ingredient" = "ingredient", "FAO_code")) %>% 
  mutate(total_crop_demand_ge = producing_ingredient_consumed_tonnes*ge_allocation_factor,
         total_crop_demand_ms = producing_ingredient_consumed_tonnes*mass_allocation_factor,
         total_crop_demand_econ = producing_ingredient_consumed_tonnes*econ_allocation_factor) %>%
  group_by(iso3c_producing, diet, fcr_type, source_ingredient, FAOSTAT_name, FAO_code) %>%
  summarise(total_crop_demand_ms = sum(total_crop_demand_ms, na.rm = TRUE),
            total_crop_demand_ge = sum(total_crop_demand_ge, na.rm = TRUE),
            total_crop_demand_econ = sum(total_crop_demand_econ, na.rm = TRUE)) %>%
  ungroup() %>%
  filter(total_crop_demand_ge != 0 & total_crop_demand_ms != 0 & total_crop_demand_econ != 0)



write.csv(crop_raw_material_demand, here("prep/02_feed/data/demand/ingredient_crop_demand_production_conversions_10-02-2024.csv"), row.names = FALSE)



```


## STEP 3 Calculate proportion of production for animal system for each country

Crop production according to GAEZ data adjusted to 2021 production. These data include all growing systems (intensive, irrigated, etc.) and are in units of metric tonnes.

```{r}

crop_production_df <- read_csv(here("prep/02_feed/data/GAEZcrop_production.csv"))
  
crop_production_global <- crop_production_df %>%
  group_by(GAEZ_category) %>%
  summarize(tonnes_global = sum(tonnes_producing_crop, na.rm=TRUE))


crop_production_global_props <- crop_production_df %>%
  left_join(crop_production_global) %>%
  mutate(prop_global = tonnes_producing_crop/tonnes_global)
```

Divide tonnes of consumption for each country/food system by total production to get the proportion of production produced for each food system. For any countries and raw materials where we allocate more raw material production than actual production, we will re-allocate the excess production proportional to global production patterns. 

For example, we allocate ~194k tonnes of rapeseed to Chile production, but Chile only produces ~160k tonnes of rapeseed crop. Given this, we will re-allocate the 194-160 = 34k tonnes of rapeseed production proportionally to global production (~24% of 34k to CAN, ~20% of 34k to CHN, ~13% of 34k to IND, etc. See `crop_production_global_props` df)


```{r}
feed_source <- read.csv(here("prep/02_feed/data/demand/ingredient_crop_demand_production_conversions_10-02-2024.csv")) %>%
  left_join(feed_cats, by = c("source_ingredient" = "ingredient", "FAOSTAT_name", "FAO_code")) %>%
  dplyr::select(-raw_material) %>%
  dplyr::rename("SPAM_super" = "map_spam_code",
                "GAEZ_category" = "gaez_code")

feed_source <- crop_raw_material_demand %>% 
    left_join(feed_cats, by = c("source_ingredient" = "ingredient", "FAOSTAT_name", "FAO_code")) %>%
  dplyr::select(-raw_material) %>%
  dplyr::rename("SPAM_super" = "map_spam_code",
                "GAEZ_category" = "gaez_code")

# if the data were perfect, none would be > 1, however, since we are modelling scenarios, we don't care if it is above 1.

proportion <- left_join(feed_source, crop_production_df, by=c("iso3c_producing", "GAEZ_category")) %>%
  mutate(tonnes_producing_crop = ifelse(is.na(tonnes_producing_crop), 0, tonnes_producing_crop)) %>%
  mutate(excess_ms = total_crop_demand_ms - tonnes_producing_crop,
         excess_ge = total_crop_demand_ge - tonnes_producing_crop,
         excess_econ = total_crop_demand_econ - tonnes_producing_crop) %>%
  mutate(excess_ms = ifelse(excess_ms < 0, 0, excess_ms),
         excess_ge = ifelse(excess_ge < 0, 0, excess_ge),
         excess_econ = ifelse(excess_econ < 0, 0, excess_econ))

excess_allocations <- proportion %>%
  group_by(diet, fcr_type, source_ingredient, FAOSTAT_name, FAO_code, GAEZ_category, SPAM_super) %>%
  summarise(excess_ms = sum(excess_ms),
            excess_ge = sum(excess_ge), 
            excess_econ = sum(excess_econ)) %>%
  ungroup() %>%
 # filter(excess_ms > 0 | excess_ge > 0 | excess_econ > 0) %>%
  right_join(crop_production_global_props) %>%
  mutate(excess_allocation_ms = excess_ms*prop_global,
         excess_allocation_ge = excess_ge*prop_global,
         excess_allocation_econ = excess_econ*prop_global) %>% 
  filter(!is.na(diet)) %>%
  dplyr::select(diet, fcr_type, iso3c_producing, source_ingredient, FAOSTAT_name, FAO_code, GAEZ_category, SPAM_super,  excess_allocation_ms, excess_allocation_ge, excess_allocation_econ) ## ok now we have the excess production re-allocated based on global production patterns. Now we need to rejoin and add to the original allocation data. 

proportion_fin = proportion %>% 
  mutate(total_crop_demand_ms = ifelse(excess_ms > 0, tonnes_producing_crop, total_crop_demand_ms),
         total_crop_demand_ge = ifelse(excess_ge > 0, tonnes_producing_crop, total_crop_demand_ge),
         total_crop_demand_econ = ifelse(excess_econ > 0, tonnes_producing_crop, total_crop_demand_econ)) %>%
  dplyr::select(-excess_ms, -excess_ge, -excess_econ) %>%
  left_join(excess_allocations) %>%
  mutate(total_crop_demand_ms = total_crop_demand_ms + excess_allocation_ms,
         total_crop_demand_ge = total_crop_demand_ge + excess_allocation_ge,
         total_crop_demand_econ = total_crop_demand_econ + excess_allocation_econ) %>%
  dplyr::select(-excess_allocation_econ, -excess_allocation_ge, -excess_allocation_ms) %>%
  mutate(prop_produced_for_system_ms = ifelse(total_crop_demand_ms == 0, 0, total_crop_demand_ms/tonnes_producing_crop),
         prop_produced_for_system_ge = ifelse(total_crop_demand_ge == 0, 0, total_crop_demand_ge/tonnes_producing_crop), 
         prop_produced_for_system_econ = ifelse(total_crop_demand_econ == 0, 0, total_crop_demand_econ/tonnes_producing_crop))



write_csv(proportion_fin, here("prep/02_feed/output/proportion_feed_per_country_system_diet_10-02-2024.csv"))



```

## Step 4

Rasterize this information to get rasters describing the amount of tonnages for each crop in each cell global that goes to salmon aquaculture feed ingredients, additionally, divide by crop yield estimates from the FAO to get estimates of crop area (km2 disturbance) in each cell that end up as salmon aquaculture feed ingredients. 

 - we will use the uncapped proportion for this for now.. we are assuming this is a scenario where we completely fulfill demand, regardless of what the current-day production looks like. So if we need to produce more crop to meet ingredient demand in a country, then we will. Nearly all of the proportions are below one anyway. 


```{r}

total_prop <- read.csv(here("prep/02_feed/output/proportion_feed_per_country_system_diet_10-02-2024.csv")) %>%
  dplyr::select(SPAM_super = SPAM_super, GAEZ_category, diet, fcr_type, ingredient = source_ingredient, iso3c_producing, total_prop_ms = prop_produced_for_system_ms, total_prop_ge = prop_produced_for_system_ge,
                total_prop_econ = prop_produced_for_system_econ) %>%
  pivot_longer(cols = c("total_prop_ms", "total_prop_ge", "total_prop_econ"), names_to = "allocation_type", values_to = "total_prop") %>%
  mutate(allocation_type = case_when(
    allocation_type == "total_prop_ge" ~ "ge",
    allocation_type == "total_prop_ms" ~ "mass",
    TRUE ~ "economic"
  ))
           

gaez_fao <- feed_cats %>%
  dplyr::distinct(FAO_code, gaez_code)

fao_yields <- readRDS(here("data/tidy_data/production-data/crops_production_tidy.rds")) %>%
  filter(element == "Yield",
         year == "2021",
         area_code < 5000,
         iso3c %in% total_prop$iso3c_producing) %>%  # filter for 2021 and yield data
  mutate(yield_tonnes_ha = value/10000) %>% # convert hg to tonnes
  left_join(gaez_fao, by = c("item_code" = "FAO_code")) %>% # get gaez codes
  filter(!is.na(gaez_code))



## organize region data
rgns <- read_csv(file.path(rdsi_raw_data_dir, "food-systems-project/food_rgns_xy.csv"))

####################
## start loop here
#################

todo <- unique(dplyr::select(total_prop, diet, fcr_type, GAEZ_category, ingredient, allocation_type))


todo <- total_prop %>%
  group_by(diet, fcr_type, GAEZ_category, ingredient, allocation_type) %>%
  summarise(sum_prop = sum(total_prop, na.rm = TRUE)) %>%
  ungroup() %>%
  filter(sum_prop >0) 


# 

for(i in 1:dim(todo)[1]){
# i=91
crop_gaez <- todo$GAEZ_category[i]
product_type <- todo$ingredient[i]
diet_type = todo$diet[i]
allocation = todo$allocation_type[i]
fcr <- todo$fcr_type[i]


## you can comment this out if you want to rerun everything
if(all(file.exists(sprintf(here("prep/02_feed/output/%s/%s/%s_%s_%s_A.tif"), diet_type, fcr, crop_gaez, product_type, allocation), sprintf(here("prep/02_feed/output/%s/%s/%s_%s_%s_P.tif"), diet_type, fcr, crop_gaez, product_type, allocation)))){
  cat("skipping because area and production rasters both exist!")
  next()
}

# total country feed production
feed_crop <- filter(total_prop, GAEZ_category==crop_gaez, diet == diet_type, fcr_type == fcr, ingredient == product_type , allocation_type == allocation) %>%
  dplyr::select(iso3c = iso3c_producing, prop_feed = total_prop)

# proportion crop grown for feed per country
prop_feed_crop <- left_join(rgns, feed_crop, by="iso3c") %>%
  dplyr::select(x, y, prop_feed)

prop_feed_crop_raster <- rast(prop_feed_crop)
# plot(prop_feed_crop_raster)
crs(prop_feed_crop_raster) <- "+proj=longlat +datum=WGS84"

tonnes_crop_raster <- rast(sprintf(file.path(gaez_dir, "scaled_prod_maps_2021/crop_%s_Total_scaled.tif"), crop_gaez))

tonnes_crop_feed_raster <- tonnes_crop_raster*prop_feed_crop_raster
# plot(log(tonnes_crop_feed_raster+1))

writeRaster(tonnes_crop_feed_raster, filename= sprintf(here("prep/02_feed/output/%s/%s/%s_%s_%s_P.tif"), diet_type, fcr, crop_gaez, product_type, allocation), overwrite=TRUE)



# country yield estimates
feed_yield <- filter(fao_yields, gaez_code == crop_gaez) %>%
  dplyr::group_by(iso3c) %>%
  summarise(yield = mean(yield_tonnes_ha, na.rm = TRUE)) %>%
  ungroup()

mean_yield <- mean(feed_yield$yield, na.rm = TRUE)

feed_yield_gf <- feed_yield %>% 
  full_join(feed_crop) %>% # join with feed crop so we can see which countries are missing yields
  filter(!is.na(prop_feed)) %>%
  mutate(yield = ifelse(is.na(yield), mean_yield, yield)) %>% # now we will gapfill the yields with the global average yield for this crop
  dplyr::select(-prop_feed)

# proportion crop grown for feed per country
yield_feed_crop <- left_join(rgns, feed_yield_gf, by="iso3c") %>%
  dplyr::select(x, y, yield)

yield_feed_crop_raster <- rast(yield_feed_crop)
# plot(yield_feed_crop_raster)

crs(yield_feed_crop_raster) <- "+proj=longlat +datum=WGS84"

area_crop_feed_raster <- (tonnes_crop_feed_raster/yield_feed_crop_raster)/100 # divide by yield to get hectares of cropland then convert hectares to km2 (divide by 100)
# plot(log(area_crop_feed_raster+1)) 
# plot(area_crop_feed_raster)

# global(area_crop_feed_raster, "sum", na.rm = TRUE) # just a check

writeRaster(area_crop_feed_raster, filename= sprintf(here("prep/02_feed/output/%s/%s/%s_%s_%s_A.tif"), diet_type, fcr, crop_gaez, product_type, allocation), overwrite=TRUE)
}

```

## Data check 

```{r}

# take a look at plant dominant
demand_test <- consumption %>%
  filter(diet == "plant-dominant", fcr_type == "regular")

sum(demand_test$consuming_ingredient_consumed_tonnes) # 3182323

test <- rast(list.files(here("prep/02_feed/output/plant-dominant/regular"), full.names = TRUE, pattern = "economic_P.tif"))
sum(global(test, "sum", na.rm = TRUE)$sum) # 3271915 - ok pretty close, that's good off by less than 100k tonnes

test_names <- str_replace_all(str_before_last(list.files(here("prep/02_feed/output/plant-dominant/regular"), pattern = "economic_A.tif"), ".tif"), "_", ":")

test <- rast(list.files(here("prep/02_feed/output/plant-dominant/regular"), full.names = TRUE, pattern = "economic_A.tif"))

names(test) <- test_names

# plot(log(test+1))

test_df <- as.data.frame(global(test, "sum", na.rm = TRUE))

test_df %>% arrange(-sum)

#                                                   sum
# Rapeseed:canola oil:economic:A             4581.70200
# Soybean:soy protein concentrate:economic:A 2459.66028
# Wheat:wheat gluten:economic:A              1909.82016
# Pulses:pea protein concentrate:economic:A   587.29734
# Pulses:faba beans:economic:A                585.16919
# Wheat:wheat:economic:A                      495.86676
# Pulses:pea flour:economic:A                 477.81140
# Sunflower:sunflower meal:economic:A         379.33434
# CropsNES:linseed oil:economic:A             323.19909
# Pulses:guar meal:economic:A                 312.98079
# Soybean:soy oil:economic:A                   73.17686
# CropsNES:coconut oil:economic:A              59.71471
# Maize:corn gluten meal:economic:A            13.26594

# take a look at fish dominant
demand_test <- consumption %>%
  filter(diet == "fish-dominant", fcr_type == "regular")

sum(demand_test$consuming_ingredient_consumed_tonnes) # 3175971 - less demand

test_names <- str_replace_all(str_before_last(list.files(here("prep/02_feed/output/fish-dominant/regular"), pattern = "economic_A.tif"), ".tif"), "_", ":")

test <- rast(list.files(here("prep/02_feed/output/fish-dominant/regular"), full.names = TRUE, pattern = "economic_A.tif"))

names(test) <- test_names

# plot(log(test+1))

test_df <- as.data.frame(global(test, "sum", na.rm = TRUE))

sum(test_df$sum) # 5114.828 km2 of cropland for salmon aquaculture feeds under fish-dominant 

test_df %>% arrange(-sum)

#                                          sum
# Wheat:wheat gluten:economic:A     2533.43490
# Soybean:soybean meal:economic:A   1563.19855
# Wheat:wheat:economic:A             534.01036
# Pulses:faba beans:economic:A       324.73856
# Maize:corn gluten meal:economic:A   75.80536


```
 
Make some plots

```{r}
feed_source <- read.csv(here("prep/02_feed/data/demand/ingredient_crop_demand_production_conversions_10-02-2024.csv")) %>%
  left_join(feed_cats, by = c("source_ingredient" = "ingredient", "FAOSTAT_name", "FAO_code")) %>%
  dplyr::select(-raw_material) %>%
  dplyr::rename("GAEZ_category" = "gaez_code", "ingredient" = "source_ingredient") %>%
  group_by(iso3c_producing, diet, fcr_type, GAEZ_category, ingredient) %>%
  summarize(produced_ingredient_tonnes_ms = sum(total_crop_demand_ms),
            produced_ingredient_tonnes_ge = sum(total_crop_demand_ge)) %>% 
  ungroup()
tmp <- filter(feed_source) %>%
    group_by(ingredient, diet, fcr_type) %>%
  slice_max(order_by = produced_ingredient_tonnes_ge, n = 15) %>%
  ungroup() %>%
  filter(produced_ingredient_tonnes_ge > 0)
ggplot(data = tmp %>%
         filter(diet == "plant-dominant"), aes(x = produced_ingredient_tonnes_ge, y = reorder_within(iso3c_producing, produced_ingredient_tonnes_ge, GAEZ_category))) + 
  geom_bar(stat = "identity") + 
  facet_grid(fcr_type~ingredient, scales = "free") + 
  labs(x = "Tonnes of crop produced for ingredient", y = "Country") + 
  theme_bw() + 
  theme(axis.text.x = element_text(angle = 30, vjust = 0.5)) 

```

Check if there are spatial differences between ingredients... I don't think there are? Obviously there are value differences. But in terms of the actual ingredient, I think we group the trade data by product code, which are usually the same. So Julia might need to report results at the raw material level rather than ingredient level. 

```{r}
## try soy protein concentrate and soy oil 
spc <- rast(here("prep/02_feed/output/plant-dominant/regular/Soybean_soy protein concentrate_economic_P.tif")) %>%
  filter(coef > 0)
soy_oil <- rast(here("prep/02_feed/output/plant-dominant/regular/Soybean_soy oil_economic_P.tif")) %>%
  filter(coef > 0)

spc
soy_oil

plot(spc)
plot(soy_oil)

spc_df <- spc %>% as.data.frame(., xy = TRUE)
soy_oil_df <- soy_oil %>% as.data.frame(., xy = TRUE)

```

```{r}
wheat_pd <-  rast(here("prep/02_feed/output/plant-dominant/regular/Wheat_wheat_economic_P.tif")) %>%
  filter(coef > 0)

wheat_fd <-  rast(here("prep/02_feed/output/fish-dominant/regular/Wheat_wheat_economic_P.tif")) %>%
  filter(coef > 0)

diff <- wheat_fd-wheat_pd 

diff[diff <=1] <- NA

plot(diff, col  = "red")

corn_pd <-  rast(here("prep/02_feed/output/plant-dominant/regular/Maize_corn gluten meal_economic_P.tif")) %>%
  filter(coef > 0)

corn_fd <-  rast(here("prep/02_feed/output/fish-dominant/regular/Maize_corn gluten meal_economic_P.tif")) %>%
  filter(coef > 0)

diff <- corn_fd-corn_pd 

diff[diff <=1] <- NA

plot(diff, col  = "red")


soy_pd <-  rast(here("prep/02_feed/output/plant-dominant/regular/Soybean_soy protein concentrate_economic_P.tif")) %>%
  filter(coef > 0)

soy_fd <-  rast(here("prep/02_feed/output/fish-dominant/regular/Soybean_soybean meal_economic_P.tif")) %>%
  filter(coef > 0)

diff <- soy_fd-soy_pd

plot(diff)
```
