==> prep/02_feed/4b_forage_fmfo_allocation_production.Rmd <==
---
title: "Forage fish for feed"
author: "Gage Clawson"
date: "May 18, 2023"
output: html_document
editor_options: 
  chunk_output_type: console
---

# Summary

GOAL: Get proportion of fish oil and fish meal live weight equivalents that is consumed by salmon aquaculture in each country per species. Allocate to live weight fish using mass, economic, and energetic allocation rates.

Outputs: a raster for each fish oil and fish meal, under each diet scenario, under each fcr scenario, describing the harvest of each FMFO live weight equivalents (tonnes) that are consumed in each country by salmon aquaculture systems.

```{r}
# load relevant packages, etc.

library(here)
library(raster)
library(tidyverse)
library(countrycode)
library(readxl)


source(here("src/fxns.R"))
source(here("src/directories.R"))

```

## Data Sources

**Reference**: Cottrell, R.S. et al. (2024) ‘Raw material origins matter as much as feed composition for the environmental footprint of aquaculture feeds’, Manuscript in review [Preprint].

**Description**: We use the raw material allocation methods created from this paper. 

**Reference**: Gaulier, G. and Zignago, S. (2010) BACI: International Trade Database at the Product-Level. The 1994-2007 Version. Working Papers 2010–23. CEPII. Available at: http://www.cepii.fr/CEPII/fr/publications/wp/abstract.asp?NoDoc=2726.

**Description**: Bilateral trade flows of fish meal and oil. 

**Reference**: Watson, R.A. (2017) ‘A database of global marine commercial, small-scale, illegal and unreported fisheries catch 1950–2014’, Scientific Data, 4(1), p. 170039. Available at: https://doi.org/10.1038/sdata.2017.39.

**Description**: Global fish catch dataset

 - See extended data tables from the publication for all species specific allocation rates used.

# Methods 

 Using BACI trade database as our trade data
  - derived from COMTRADE
  - README: http://www.cepii.fr/DATA_DOWNLOAD/baci/doc/DescriptionBACI.html
  - BACI provides yearly data on bilateral trade flows at the product level. Products are identified using the Harmonized System (HS), which is the standard nomenclature for international trade, used by most customs. 
  - We filter for HS codes 230120 (fish meals) and 150420 (fish oils)
  - Using year 2017 so we match with Watson catch data 
  
https://hstracker.wto.org/?_inputs_&sidebarCollapsed=false&page=%22visualizer%22&hscode=%22150420%22
  
```{r}
baci_files <- list.files(file.path(rdsi_raw_data_dir, "BACI_data/BACI_HS12_V202301/"), full.names = TRUE)[1:8]

baci_raw <- lapply(baci_files, read.csv)


baci_list_filter <- map(baci_raw, ~ .x[.x$k %in% c(230120, 150420), ])

baci_raw_df <- bind_rows(baci_list_filter)

colnames(baci_raw_df)

# i = importer country code? 
# j = partner country code? 
# k = product code
# t = year
# v = value (USD?)
# q = quantity

product_codes <- read.csv(file.path(rdsi_raw_data_dir, "BACI_data/BACI_HS12_V202301/product_codes_HS12_V202301.csv"))

country_codes <- read.csv(file.path(rdsi_raw_data_dir, "BACI_data/BACI_HS12_V202301/country_codes_V202301.csv"))

baci_filt <- baci_raw_df %>% 
  left_join(product_codes, by = c("k" = "code")) %>%
  rename(year = t, importer_code = j, partner_code = i, code = k) %>%
  left_join(country_codes, by = c("importer_code" = "country_code")) %>%
  dplyr::select(year, importer_code, importer_name = country_name_full, importer_iso3c = iso_3digit_alpha, partner_code, hs_code = code, v, q, description) %>%
    left_join(country_codes, by = c("partner_code" = "country_code")) %>%
  dplyr::select(year, importer_code, importer_name , importer_iso3c, partner_code, partner_name = country_name_full, partner_iso3c = iso_3digit_alpha, hs_code, v, q, description) %>%
  filter(!is.na(partner_iso3c)) %>%
    filter(!is.na(importer_iso3c)) %>%
    filter(partner_iso3c != "N/A") %>%
    filter(importer_iso3c != "N/A") %>%
  mutate(q = as.numeric(q)) %>%
  filter(!is.na(q)) %>%
  group_by(year, importer_name, importer_iso3c, partner_name, partner_iso3c, hs_code, description) %>% 
  summarise(qty = sum(q, na.rm = TRUE)) %>%
  ungroup() %>% 
    group_by(importer_name, importer_iso3c, partner_name, partner_iso3c, hs_code, description) %>% 
  summarise(qty = mean(qty, na.rm = TRUE)) %>%
  ungroup()

test <- baci_filt %>%
  filter(importer_iso3c == partner_iso3c) # no domestic?


test <- baci_filt %>%
  group_by(partner_iso3c, hs_code) %>%
  summarise(qty = sum(qty, na.rm = TRUE)) %>%
  ungroup()
  

write.csv(baci_filt, file.path(rdsi_raw_data_dir, "BACI_data/BACI_FMFO_codes.csv"), row.names = FALSE)

```


## Approach 

 1. Start with demand per country 
    - EX: NOR requires X tonnes of FO
 2. Get trade data for FO and FM and match to demand data. Figure out how much countries import FO or FM from other countries. 
    - EX: NOR imports Y% of FO from USA, Z% from Peru, etc.
 3. Multiply the import %s by the demand tonnes to get ingredient tonnes imported per country
    - EX: NOR requires 100 tonnes of FO, they import 60% from the USA, then 60 tonnes of FO ingredients in NOR comes from the USA for salmon. 
 4. Group by exporter country to get total amount of FO or FM ingredients they export under each diet demand scenario
    - EX: USA exports 60 tonnes to NOR, 40 tonnes to CHL, then they export 100 tonnes of FO total for salmon feed. 
 5. Determine each countries' breakdown of forage fish species catch by flag country (e.g., where the catch is actually landed, rather than caught). Filter for all forage fish species, group by country and species, and determine the within-country proportions of catch from each species. For example, 55% of the forage fish species weight that USA lands is from Gulf Menhaden, 24% from Atlantic Menhaden, 8% from Atka mackerel, etc...
 6. Apply these within country percentages to total country exported fish oil and fish meal (calculated in step 4). Now we will have X tonnes of fish oil and Y tonnes of fish meal exported from or kept within country Z per fish species per diet scenario. 
     - For example, if 55% of the forage catch in USA is from Gulf Menhaden, then 55% of the fish meal (and 55% of the fish oil) exported from or kept in USA will be attributed to Gulf Menhaden. 
 7. Next, we can use the species specific allocation rates that Rich put together to convert the fish oil and fish meal to raw material live weight equivalents. Now we have total tonnes of fish oil and fish meal live weight equivalents per landing country per species. 
    - for adjustments: 
       - If possible use species and region specific rates 
       - If a species exists outside of a region, use global species specific rates 
       - Otherwise, use a global average of all species and regions 

 8. Divide the tonnes of fish oil and fish meal live weight equivalents by the total amount of forage catch in each country. This gives us a proportion of forage fish catch in each country that goes to FMFO. 
 9. Join with Watson data, group by cell, and rasterize. Now we will have rasters describing the amount of forage fish catch that goes to fish oil or fish meal for salmon feed under each diet scenario under each allocation approach (economic, mass, or energetic). 
 
 
### step 1 
 
 Start with demand per country 
    - EX: NOR requires X tonnes of FO
    
```{r}
## step 1
country_demand <- read_csv(here("prep/02_feed/data/demand/total_aquaculture_ingredient_demand.csv")) %>%
  filter(ingredient %in% c("fish meal, forage fish", "fish oil, forage fish")) %>%
  mutate(ingredient = ifelse(ingredient == "fish meal, forage fish", "fish meal", "fish oil")) %>%
    pivot_longer(cols = c("ingredient_demand", "ingredient_demand_efficient"), names_to = "fcr_type", values_to = "demand") %>%
  mutate(fcr_type = ifelse(fcr_type == "ingredient_demand", "regular", "efficient")) %>% 
  select(iso3c, animal, diet, ingredient, fcr_type, ingredient_demand = demand) 

sum(country_demand$ingredient_demand) # 3565490/2 # checks out

```



### step 2

 2. Get trade data for FO and FM and match to demand data. Figure out how much countries import FO or FM from other countries. 
    - EX: NOR imports Y% of FO from USA, Z% from Peru, etc.
    
```{r}
baci_filt <- read.csv(file.path(rdsi_raw_data_dir, "BACI_data/BACI_FMFO_codes.csv"))

oils_meals_import_props <- baci_filt %>% 
  mutate(type = ifelse(hs_code == 230120, "fish meal", "fish oil")) %>%
  dplyr::select(importer_iso3c, partner_iso3c, type, qty) 
  
length(unique(oils_meals_import_props$partner_iso3c)) # 164 countries...


## Since Faroe Islands are missing, lets duplicate the Iceland trading props, and use those for Faroe Islands. I think this is defensible. Geographically they are similar. Both heavily rely on marine resources as products. Both a part of the European Free Trade Association. 

iceland_faroe_props <- oils_meals_import_props %>%
  filter(importer_iso3c == "ISL" | partner_iso3c == "ISL") %>%
  mutate(importer_iso3c = ifelse(importer_iso3c == "ISL", "FRO", importer_iso3c),
         partner_iso3c = ifelse(partner_iso3c == "ISL", "FRO", partner_iso3c))

oils_meals_import_props <- rbind(oils_meals_import_props, iceland_faroe_props) %>%
    group_by(importer_iso3c, type) %>%
  mutate(total_import = sum(qty, na.rm = TRUE)) %>% 
  ungroup() %>%
  mutate(prop_import = qty/total_import)

```

### step 3

Multiply the import %s by the demand tonnes to get ingredient tonnes imported per country
    - EX: NOR requires 100 tonnes of FO, they import 60% from the USA, then 60 tonnes of FO ingredients in NOR comes from the USA for salmon. 

```{r}
## step 3

demand_trade <- country_demand %>%
  left_join(oils_meals_import_props, by = c("iso3c" = "importer_iso3c", "ingredient" = "type")) %>% 
  mutate(tonnes_ingredient_demand_traded = prop_import*ingredient_demand) 

unique(demand_trade$iso3c)
length(unique(demand_trade$partner_iso3c)) # 113 countries

sum(demand_trade$tonnes_ingredient_demand_traded, na.rm = TRUE) # 3565490/2 # looks good
```

### Step 4

Group by exporter country to get total amount of FO or FM ingredients they export under each diet demand scenario
    - EX: USA exports 60 tonnes to NOR, 40 tonnes to CHL, then they export 100 tonnes of FO total for salmon feed.
    
```{r}

## step 4
demand_trade_origin <- demand_trade %>%
  group_by(partner_iso3c, ingredient, diet, fcr_type) %>%
  summarise(total_ingredient_tonnes_traded = sum(tonnes_ingredient_demand_traded, na.rm = TRUE)) %>% 
  ungroup()

sum(demand_trade_origin$total_ingredient_tonnes_traded) # 3565490/2

```

### Step 5 

Determine each countries' breakdown of forage fish species catch by flag country (e.g., where the catch is actually landed, rather than caught). Filter for all forage fish species, group by country and species, and determine the within-country proportions of catch from each species. For example, 55% of the forage fish species weight that USA lands is from Gulf Menhaden, 24% from Atlantic Menhaden, 8% from Atka mackerel, etc...
 
```{r}
## step 5 

gear_df <- read.csv(file.path(watson_dir, "v5.0/Codes_gear.csv")) %>%
  janitor::clean_names()

gear_clean_df <- gear_df %>%
  dplyr::select(gear, f_gear_code, f_gear_label, fleet_gear_name) %>%
  distinct() %>%
   mutate(f_gear_label = tolower(f_gear_label),
         fleet_gear_name = tolower(fleet_gear_name),
         fleet_gear_name = str_replace(fleet_gear_name, 'tuna', ' tuna'),
         fleet_gear_name = str_replace_all(fleet_gear_name, '[^a-z]+', '_'))

benthic_gears <- c('trawl', 'dredge', 'trap')

watson_raw <- read_csv(file.path(watson_dir, "v5.0/watson_2017_catch_w_forage_id.csv"))

watson_cells <- read.csv(file.path(watson_dir, "v5.0/Codes_cells.csv"))

watson_rgn <- watson_raw %>%
  distinct(CNumber, country = CountryName) %>%
  mutate(iso3c = countrycode(sourcevar = country, origin = "country.name", destination = "iso3c")) %>%
  mutate(iso3c = 
           case_when(
             country == "Amer Samoa" ~ "ASM", 
             country == "Br Ind Oc Tr" ~ "IOT", 
             country == "Br Virgin Is" ~ "VGB", 
             country == "Dominican Rp" ~ "DOM",
             country == "Fr Polynesia" ~ "PYF", 
             country == "Fr Guiana" ~ "GUF", 
             country == "Micronesia" ~ "FSM", 
             country == "St Pier Mq" ~ "SPM", 
             country == "Untd Arab Em" ~ "UAE", 
             country == "US Virgin Is" ~ "VIR", 
             TRUE ~ iso3c
           ))


codes_country <- read.csv(file.path(watson_dir, "v5.0/Codes_country.csv")) # read in country codes 


# calculate forage catch and add iso3c codes
forage_catch <- 
  watson_raw |> 
  dplyr::select(-foragefish) %>% 
  left_join(watson_rgn, by = c("CountryName" = "country", "CNumber")) %>%
  mutate(catch = ReportedIND  + ReportedNIND) %>%
  mutate(forage_id = ifelse(CommonName == "Antarctic krill", 1, forage_id)) %>%
  mutate(FOFM_catch = catch*forage_id) %>%
  filter(!is.na(FOFM_catch)) %>% 
  left_join(gear_clean_df, by = c("Gear" = "gear", "FGearCode"= "f_gear_code"))

foragefish <-  # we need to make a dataframe that has all of the catch in it.. so we can get later get a prop of forage catch to total catch. 
  watson_raw |> 
  dplyr::select(-foragefish) %>%
  left_join(watson_rgn, by = c("CountryName" = "country", "CNumber")) %>%
    mutate(forage_id = ifelse(CommonName == "Antarctic krill", 1, forage_id)) %>%
  mutate(catch = ReportedIND + ReportedNIND) %>%
  mutate(FOFM_catch = catch*forage_id) %>%
  mutate(FOFM_catch = ifelse(is.na(FOFM_catch), 0, FOFM_catch)) %>%
  left_join(gear_clean_df, by = c("Gear" = "gear", "FGearCode"= "f_gear_code")) %>%
  mutate(water_col_position = ifelse(fleet_gear_name %in% benthic_gears, "benthic", "pelagic")) %>%
  mutate(prop_discards = (DiscardsIND + DiscardsNIND)/(DiscardsIND + DiscardsNIND + FOFM_catch)) %>%
  mutate(prop_discards = ifelse(FOFM_catch == 0, 0, prop_discards))

write_rds(foragefish, file.path(watson_dir, "v5.0/int/all_catch_foragefish.rds"))

## get an estimate of proportion of forage fish catch coming from different species in different landing countries. 
forage_catch_iso <- forage_catch %>%
  group_by(iso3c, CommonName, TaxonName) %>%
  summarise(catch = sum(FOFM_catch, na.rm = TRUE)) %>%
  ungroup() %>%
  group_by(iso3c) %>%
  mutate(total_fofm_catch = sum(catch, na.rm = TRUE)) %>%
  ungroup() %>%
  mutate(prop_of_iso_total = catch/total_fofm_catch)

forage_catch_iso_no_krill <- forage_catch %>% ## we need one without krill for fish oil, since krill isn't included for fish oil. 
  filter(CommonName != "Antarctic krill")  %>%
    group_by(iso3c, CommonName, TaxonName) %>%
  summarise(catch = sum(FOFM_catch, na.rm = TRUE)) %>%
  ungroup() %>%
  group_by(iso3c) %>%
  mutate(total_fofm_catch = sum(catch, na.rm = TRUE)) %>%
  ungroup() %>%
  mutate(prop_of_iso_total = catch/total_fofm_catch)

```


### step 6 

Apply these within country percentages to total country exported fish oil and fish meal (calculated in step 4). Now we will have X tonnes of fish oil and Y tonnes of fish meal exported from or kept within country Z per fish species per diet scenario. 
     - For example, if 55% of the forage catch in USA is from Gulf Menhaden, then 55% of the fish meal (and 55% of the fish oil) exported from or kept in USA will be attributed to Gulf Menhaden. 
     
```{r}
## step 6 

forage_fofm_origins_meal <- forage_catch_iso %>%
  dplyr::select(iso3c, CommonName, TaxonName, prop_of_iso_total) %>%
  left_join(demand_trade_origin %>% filter(ingredient == "fish meal"), by = c("iso3c" = "partner_iso3c")) %>%
  mutate(tonnes_fofm_spp = total_ingredient_tonnes_traded*prop_of_iso_total) %>%
  dplyr::select(iso3c, CommonName, TaxonName, ingredient, diet, fcr_type, tonnes_fofm_spp) %>%
  filter(!is.na(tonnes_fofm_spp))

test <- forage_fofm_origins_meal %>%
  pivot_wider(names_from = fcr_type, values_from = tonnes_fofm_spp) %>%
  mutate(diff = (regular - efficient)/regular) # ok so good here

forage_fofm_origins_oil <- forage_catch_iso_no_krill %>%
  dplyr::select(iso3c, CommonName, TaxonName, prop_of_iso_total) %>%
  left_join(demand_trade_origin %>% filter(ingredient == "fish oil"), by = c("iso3c" = "partner_iso3c")) %>%
  mutate(tonnes_fofm_spp = total_ingredient_tonnes_traded*prop_of_iso_total) %>%
  dplyr::select(iso3c, CommonName, TaxonName, ingredient, diet, fcr_type, tonnes_fofm_spp) %>%
  filter(!is.na(tonnes_fofm_spp))


test <- forage_fofm_origins_oil %>%
  pivot_wider(names_from = fcr_type, values_from = tonnes_fofm_spp) %>%
  mutate(diff = (regular - efficient)/regular) # ok so good here

forage_fofm_origins <- rbind(forage_fofm_origins_meal, forage_fofm_origins_oil)

sum(forage_fofm_origins$tonnes_fofm_spp)
# 3527008/2

sum(country_demand$ingredient_demand) # 3565490 missing ~38k tonnes... 

```


### step 7 

Next, we can use the species specific allocation rates that Rich put together to convert the fish oil and fish meal to raw material live weight equivalents. Now we have total tonnes of fish oil and fish meal live weight equivalents per landing country per species. 
    - for adjustments: 
       - If possible use species and region specific rates 
       - If a species exists outside of a region, but is in the allocation data, use global species specific rates. right now we only have species and region specific allocation rates. so this is kind of moot. 
       - Otherwise, use a global average of all species and regions 

```{r}

## step 7

## read in allocation data 
fish_allocations_raw <- read_xlsx(here("data/tidy_data/allocation/embodied_fish_allocation.xlsx"))

# GA Global average, SO Southern Ocean, NS North Sea, IS Icelandic Shelf, NW Norwegian Sea, BS Barents Sea, HC Humboldt Current, GM Gulf of Mexico.

ga_mass_FO <- fish_allocations_raw %>%
  filter(ecosystem == "GA", ingredient == "Fish oil") %>%
  pull(mass_value)

ga_mass_FM <- fish_allocations_raw %>%
  filter(ecosystem == "GA", ingredient == "Fishmeal") %>%
  pull(mass_value)

  
ga_ge_FO <- fish_allocations_raw %>%
  filter(ecosystem == "GA", ingredient == "Fish oil") %>%
  pull(ge_value)


ga_ge_FM <- fish_allocations_raw %>%
  filter(ecosystem == "GA", ingredient == "Fishmeal") %>%
  pull(ge_value)


ga_econ_FO <- fish_allocations_raw %>%
  filter(ecosystem == "GA", ingredient == "Fish oil") %>%
  pull(econ_value)


ga_econ_FM <- fish_allocations_raw %>%
  filter(ecosystem == "GA", ingredient == "Fishmeal") %>%
  pull(econ_value)

fish_allocations_tidy <- fish_allocations_raw %>%
  filter(!is.na(fishmeal_yield)) %>%
    filter(ecosystem != "GA") %>%
    mutate(ecosystem_combined = case_when(grepl("IS|NWS|NS|BS", ecosystem) ~ "North Atlantic",
                                        ecosystem == "HC" ~ "Humboldt",
                                        ecosystem == "GM" ~ "Gulf of Mexico", 
                                        ecosystem == "SO" ~ "Southern Ocean",   
                                        ecosystem == "GA" ~ "Other")) %>%
  mutate(ingredient = ifelse(ingredient == "Fishmeal", "fish meal", "fish oil"))
  


fish_allocations_df_spp_rgn <- fish_allocations_tidy %>%
  group_by(CommonName, sci_name, ingredient, ecosystem_combined) %>%
  summarise(ge_value_spp_rgn = mean(ge_value, na.rm = TRUE), 
            mass_value_spp_rgn = mean(mass_value, na.rm = TRUE),
            econ_value_spp_rgn = mean(econ_value, na.rm = TRUE)) %>%
  ungroup()


fish_allocations_df_spp <- fish_allocations_tidy %>%  
  group_by(CommonName, sci_name, ingredient) %>%
  summarise(ge_value_spp = mean(ge_value),
         mass_value_spp = mean(mass_value),
         econ_value_spp = mean(econ_value)) %>%
  ungroup()


fish_allocations_df <- fish_allocations_tidy %>%
  group_by(CommonName, sci_name, ingredient) %>% # lets just do global averages per species for now
  summarise(ge_value = mean(ge_value, na.rm = TRUE),
            mass_value = mean(mass_value, na.rm = TRUE),
            econ_value = mean(econ_value, na.rm = TRUE)) %>%
  ungroup() 

## Match allocation data to the fish oil and fish meal per species from comtrade

forage_fofm_allocation <- forage_fofm_origins %>%
  left_join(fish_allocations_df_spp_rgn, by = c("ingredient", "CommonName", "TaxonName" = "sci_name")) %>%
  left_join(fish_allocations_df_spp, by = c("ingredient", "CommonName", "TaxonName" = "sci_name")) %>% 
  mutate(ge_value = case_when(
    is.na(ge_value_spp_rgn) & ingredient == "fish meal" ~ ga_ge_FM,
    is.na(ge_value_spp_rgn) & ingredient == "fish oil" ~ ga_ge_FO, 
    TRUE ~ ge_value_spp_rgn),
    mass_value = case_when(
      is.na(mass_value_spp_rgn) & ingredient == "fish meal" ~ ga_mass_FM,
      is.na(mass_value_spp_rgn) & ingredient == "fish oil" ~ ga_mass_FO, 
      TRUE ~ mass_value_spp_rgn
    ),
    econ_value = case_when(
      is.na(econ_value_spp_rgn) & ingredient == "fish meal" ~ ga_econ_FM,
      is.na(econ_value_spp_rgn) & ingredient == "fish oil" ~ ga_econ_FO, 
      TRUE ~ econ_value_spp_rgn
    )
  ) %>%
  pivot_longer(cols = c("ge_value", "mass_value", "econ_value"), 
               names_to = "allocation_type", 
               values_to = "allocation_value"
               ) %>%
  mutate(allocation_type = case_when(
    allocation_type == "ge_value" ~ "ge",
    allocation_type == "mass_value" ~ "mass",
    TRUE ~ "economic"
  )) %>%
  mutate(tonnes_fofm_spp_live_weight = tonnes_fofm_spp*allocation_value) %>%
  dplyr::select(iso3c, CommonName, TaxonName, ingredient, diet, fcr_type, tonnes_fofm_spp, allocation_type, allocation_value, tonnes_fofm_spp_live_weight)

sum(forage_fofm_allocation$tonnes_fofm_spp_live_weight) # 41159607/3 = 13719869 seems excessive, ~3.8x more than the demand. Do fish weigh 4x more than meal and oil? 

13719869/3484838 # 3.853991

mean(forage_fofm_allocation$allocation_value) # 3.879534 yeah, they do. 


test <- forage_fofm_allocation %>%
  pivot_wider(names_from = fcr_type, values_from = tonnes_fofm_spp_live_weight) %>%
  mutate(diff = (regular - efficient)/regular) # ok so good here


```


### step 8

Divide the tonnes of fish oil and fish meal live weight equivalents by the total amount of forage catch in each country. This gives us a proportion of forage fish catch in each country that goes to FMFO. 

```{r}

## step 8

# Divide the tonnes of fish oil and fish meal live weight equivalents by the total amount of forage catch in each country. This gives us a proportion of forage fish catch in each country that goes to FMFO.

foragefish_rgn_totals_meal <- forage_catch_iso %>%
  group_by(iso3c, TaxonName) %>%
  summarise(FOFM_catch = sum(catch)) %>%
  ungroup() %>%
  mutate(ingredient = "fish meal")

foragefish_rgn_totals_oil <- forage_catch_iso_no_krill %>%
  group_by(iso3c, TaxonName) %>%
  summarise(FOFM_catch = sum(catch)) %>%
  ungroup() %>%
  mutate(ingredient = "fish oil")

foragefish_rgn_totals <- rbind(foragefish_rgn_totals_meal, foragefish_rgn_totals_oil)

forage_fofm_allocation_totals <- forage_fofm_allocation %>%
  group_by(iso3c, ingredient, diet, fcr_type, TaxonName, allocation_type) %>%
  summarise(total_tonnes_fofm_lw = sum(tonnes_fofm_spp_live_weight, na.rm = TRUE)) %>%
              ungroup()

FOFM_props <- left_join(foragefish_rgn_totals, forage_fofm_allocation_totals) %>%
  filter(!is.na(diet)) %>% # filter out NAs since these countries dont produce any fofm catch for salmon feeds
  mutate(prop = ifelse(FOFM_catch == 0, 0, total_tonnes_fofm_lw/FOFM_catch)) %>%
  dplyr::select(iso3c, ingredient, diet, fcr_type, TaxonName, allocation_type, prop) %>%
  filter(!is.na(iso3c)) ## SAVE THIS FILE AND IN NEXT SCRIPT WE CAN ADD NPP ADJUSTMENT AND SAVE RASTERS

write_rds(FOFM_props, here("prep/02_feed/data/embodied_FMFO_forage_rgns.rds"))

FOFM_props <- readRDS(here("prep/02_feed/data/embodied_FMFO_forage_rgns.rds"))

foragefish <- readRDS(file.path(watson_dir, "v5.0/int/all_catch_foragefish.rds"))
watson_cells <- read.csv(file.path(watson_dir, "v5.0/Codes_cells.csv"))


diets <- unique(FOFM_props$diet)
allocations <- unique(FOFM_props$allocation_type)
ingredients <- unique(FOFM_props$ingredient)
depth_positions <- c("benthic", "pelagic")
fcrs <- unique(FOFM_props$fcr_type)

for(diet_type in diets){
  for(fcr in fcrs){
  for(allocation in allocations){
    for(ingredient_type in ingredients){
      for(pos in depth_positions){

# diet_type = "fish-dominant"
# allocation = "economic"
# ingredient_type = "fish meal"
#  pos = "benthic"
#  fcr = "efficient"

        ## feel free to comment this out if you want to rerun things
        if(all(file.exists(sprintf(here("prep/02_feed/output/%s/%s/forage fish_%s_%s_%s_bycatch_P.tif"), diet_type, fcr, ingredient_type, allocation, pos), sprintf(here("prep/02_feed/output/%s/%s/forage fish_%s_%s_%s_catch_P.tif"), diet_type, fcr, ingredient_type, allocation, pos)))){

          print(cat("rasters exists, skipping!"))

          next()
        }
        
  watson_raster_template <- terra::rast(ncol=720, nrow=360, vals=c(1:259200))

  system_prop <- FOFM_props %>%
    dplyr::filter(diet == diet_type,
                  allocation_type == allocation,
                  ingredient == ingredient_type, 
                  fcr_type == fcr) %>%
    dplyr::select(iso3c, diet, fcr_type, TaxonName, ingredient, prop)
  
  ## first do regular catch
  
  foragefish_system_catch <- left_join(foragefish, system_prop, by = c("iso3c", "TaxonName")) %>%
    filter(!is.na(ingredient)) %>%
    mutate(prop = ifelse(is.na(prop), 0, prop)) %>%
    mutate(FOFM_catch_foodsystem = prop*FOFM_catch) %>% 
    filter(water_col_position == pos) %>% ## filter for DEPTH RANGE HERE
    group_by(Cell) %>%
    summarise(tonnes = sum(FOFM_catch_foodsystem, na.rm = TRUE)) %>%
    ungroup() %>%
    left_join(watson_cells) %>%
    dplyr::select(LonCentre, LatCentre, tonnes)

  
  forage_fish_system_catch <- rast(foragefish_system_catch, type = "xyz", crs = crs(watson_raster_template)) %>%
    project(watson_raster_template)

# plot(log(forage_fish_system_catch + 1))

    writeRaster(forage_fish_system_catch, sprintf(here("prep/02_feed/output/%s/%s/forage fish_%s_%s_%s_catch_P.tif"), diet_type, fcr, ingredient_type, allocation, pos), overwrite = TRUE)
    
    ## now do bycatch 
    foragefish_system_bycatch <- left_join(foragefish, system_prop, by = c("iso3c", "TaxonName")) %>%
    filter(!is.na(ingredient)) %>%
    mutate(prop = ifelse(is.na(prop), 0, prop)) %>%
    mutate(FOFM_catch_foodsystem = prop*FOFM_catch*prop_discards) %>% 
    filter(water_col_position == pos) %>% ## filter for DEPTH RANGE HERE
    group_by(Cell) %>%
    summarise(tonnes = sum(FOFM_catch_foodsystem, na.rm = TRUE)) %>%
    ungroup() %>%
    left_join(watson_cells) %>%
    dplyr::select(LonCentre, LatCentre, tonnes)

  
  forage_fish_system_bycatch <- rast(foragefish_system_bycatch, type = "xyz", crs = crs(watson_raster_template)) %>%
    project(watson_raster_template)

# plot(log(forage_fish_system_bycatch + 1))
  print(global(forage_fish_system_bycatch, "sum", na.rm = TRUE)/global(forage_fish_system_catch, "sum", na.rm = TRUE))

    writeRaster(forage_fish_system_bycatch, sprintf(here("prep/02_feed/output/%s/%s/forage fish_%s_%s_%s_bycatch_P.tif"), diet_type, fcr, ingredient_type, allocation, pos), overwrite = TRUE)

      }
    }
   }
  }
}

```

## Data check 

```{r}
forage_files <- list.files(here("prep/02_feed/output/plant-dominant/regular/"), pattern = "bycatch_P|catch_P", full.names = TRUE)
forage_files <- forage_files[grep("forage fish", forage_files)]

new_rasts <- rast(forage_files)

new_names <- basename(forage_files)

names(new_rasts) <- new_names

# plot(log(new_rasts+1))

data.frame(test <- global(new_rasts, "sum", na.rm = TRUE)) %>% sum() #  6575817/3 # 2191939 lines up pretty well

```
