==> prep/02_feed/4c_trimmings_fmfo_allocation_production.Rmd <==
---
title: "Trimmings for feed"
author: "Gage Clawson (IMAS)"
date: "May 25, 2023"
output: html_document
editor_options: 
  chunk_output_type: console
---

# Summary 

GOAL: Get proportion of fish oil and fish meal live weight equivalents that is consumed by salmon aquaculture in each country per species that come from fish trimmings. Allocate to live weight fish using mass, economic, and energetic allocation rates. 

Outputs: a raster and dataset for each fish oil and fish meal, under each diet scenario, describing the harvest of each FMFO live weight equivalents (tonnes) from trimmings that are consumed in each country by salmon aquaculture systems.

## Data Sources

**Reference**: Cottrell, R.S. et al. (2024) ‘Raw material origins matter as much as feed composition for the environmental footprint of aquaculture feeds’, Manuscript in review [Preprint].

**Description**: We use the raw material allocation methods created from this paper. 

**Reference**: Gaulier, G. and Zignago, S. (2010) BACI: International Trade Database at the Product-Level. The 1994-2007 Version. Working Papers 2010–23. CEPII. Available at: http://www.cepii.fr/CEPII/fr/publications/wp/abstract.asp?NoDoc=2726.

**Description**: Bilateral trade flows of fish meal and oil. 

**Reference**: Watson, R.A. (2017) ‘A database of global marine commercial, small-scale, illegal and unreported fisheries catch 1950–2014’, Scientific Data, 4(1), p. 170039. Available at: https://doi.org/10.1038/sdata.2017.39.

**Description**: Global fish catch dataset

**Reference**: BioMar Group Integrated Sustainability Report (2023). BioMar Group, p. 56. Available at: https://www.biomar.com/our-promise/sustainability-report.


 - See extended data tables from the publication for all species specific allocation rates used.

```{r}
# load relevant packages, etc.

library(here)
library(raster)
library(tidyverse)
library(countrycode)
library(readxl)


source(here("src/fxns.R"))
source(here("src/directories.R"))



watson_raw <- read_csv(file.path(watson_dir, "v5.0/watson_2017_fisheries_catch_w_cat.csv"))


watson_rgn <- watson_raw %>%
  distinct(CNumber, country = CountryName) %>%
  mutate(iso3c = countrycode(sourcevar = country, origin = "country.name", destination = "iso3c")) %>%
  mutate(iso3c = 
           case_when(
             country == "Amer Samoa" ~ "ASM", 
             country == "Br Ind Oc Tr" ~ "IOT", 
             country == "Br Virgin Is" ~ "VGB", 
             country == "Dominican Rp" ~ "DOM",
             country == "Fr Polynesia" ~ "PYF", 
             country == "Fr Guiana" ~ "GUF", 
             country == "Micronesia" ~ "FSM", 
             country == "St Pier Mq" ~ "SPM", 
             country == "Untd Arab Em" ~ "UAE", 
             country == "US Virgin Is" ~ "VIR", 
             TRUE ~ iso3c
           ))


```


# Methods 

 1. Start with demand per country 
    - EX: NOR requires X tonnes of FO
 2. Get trade data for FO trimmings and FM from trimmings and match to demand data. Figure out how much countries import FO or FM from other countries. 
    - EX: NOR imports Y% of FO trimmings from USA, Z% from Peru, etc.
 3. Multiply the import %s by the demand tonnes to get ingredient tonnes imported per country
    - EX: NOR requires 100 tonnes of FO from trimmings, they import 60% from the USA, then 60 tonnes of FO ingredients in NOR comes from the USA for salmon. 
 4. Group by exporter country to get total amount of FO or FM ingredients from trimmings they export under each diet demand scenario
    - EX: USA exports 60 tonnes to NOR, 40 tonnes to CHL, then they export 100 tonnes of FO from trimmings total for salmon feed. 
 5. Determine each countries' breakdown of trimmings species catch by flag country (e.g., where the catch is actually landed, rather than caught). Filter for all trimmings species (from biomar report), group by country and species, and determine the within-country proportions of catch from each species. For example, 55% of the trimmings species weight that USA lands is from XX spp, 24% from YY spp, etc...
 6. Apply these within country percentages to total country exported fish oil and fish meal (calculated in step 4). Now we will have X tonnes of fish oil and Y tonnes of fish meal exported from or kept within country Z per trimmings fish species per diet scenario. 
     - For example, if 55% of the trimmings catch in USA is from XX spp, then 55% of the fish meal (and 55% of the fish oil) exported from or kept in USA will be attributed to XX spp. 
 7. Next, we can use species specific allocation rates to convert the fish oil and fish meal from trimmings to raw material live weight equivalents. Now we have total tonnes of fish oil and fish meal live weight equivalents per landing country per species. 
    - for adjustments: 
       - If possible use species and region specific rates 
       - If a species exists outside of a region, use global species specific rates 
       - Otherwise, use a global average of all species and regions 

 8. Divide the tonnes of fish oil and fish meal live weight equivalents by the total amount of trimmings catch in each country. This gives us a proportion of trimmings fish catch in each country that goes to FMFO. 
 9. Join with Watson data, group by cell, and rasterize? Now we will have rasters describing the amount of trimmings fish catch that goes to fish oil or fish meal for salmon feed under each diet scenario under each allocation approach (mass or energetic). 
 
 
### step 1 
 
 Start with demand per country 
    - EX: NOR requires X tonnes of FO from trimmings
    
```{r}


## step 1
country_demand <- read_csv(here("prep/02_feed/data/demand/total_aquaculture_ingredient_demand.csv")) %>%
  filter(ingredient %in% c("fish meal, cut offs", "fish oil, cut offs")) %>% 
      pivot_longer(cols = c("ingredient_demand", "ingredient_demand_efficient"), names_to = "fcr_type", values_to = "demand") %>%
  mutate(fcr_type = ifelse(fcr_type == "ingredient_demand", "regular", "efficient")) %>% 
  select(iso3c, animal, diet, ingredient, fcr_type, ingredient_demand = demand) 


forage_demand <- read_csv(here("prep/02_feed/data/demand/total_aquaculture_ingredient_demand.csv")) %>%
  filter(ingredient %in% c("fish meal, forage fish", "fish oil, forage fish")) %>% 
      pivot_longer(cols = c("ingredient_demand", "ingredient_demand_efficient"), names_to = "fcr_type", values_to = "demand") %>%
  mutate(fcr_type = ifelse(fcr_type == "ingredient_demand", "regular", "efficient")) %>% 
  select(iso3c, animal, diet, ingredient, fcr_type, ingredient_demand = demand) 




sum(country_demand$ingredient_demand)/2 # 


sum(forage_demand$ingredient_demand, na.rm = TRUE)/2 #

1105093/3484838 # 0.3171146 trimmings demand represent ~31% of all FMFO in diet

5.3/(8.3+8.8) # 0.3099415 good, matches the Aas paper. So nothing wrong there... 

```



### step 2

 2. Get trade data for FO trimmings and FM from trimmings and match to demand data. Figure out how much countries import FO or FM from other countries. 
    - EX: NOR imports Y% of FO trimmings from USA, Z% from Peru, etc.
    
```{r}
oils_meals_all <- read.csv(file.path(rdsi_raw_data_dir, "BACI_data/BACI_FMFO_codes.csv"))

oils_meals_import_props <- oils_meals_all %>% 
  mutate(type = ifelse(hs_code == 230120, "fish meal", "fish oil")) %>%
  dplyr::select(importer_iso3c, partner_iso3c, type, qty) %>%
  filter(partner_iso3c %in% c(watson_rgn$iso3c)) # filter out any presumably landlocked countries... 
  
length(unique(oils_meals_import_props$partner_iso3c)) # 135 countries...


## Since Faroe Islands are missing, lets duplicate the Iceland trading props, and use those for Faroe Islands. I think this is defensible. Geographically they are similar. Both heavily rely on marine resources as products. Both a part of the European Free Trade Association. 

iceland_faroe_props <- oils_meals_import_props %>%
  filter(importer_iso3c == "ISL" | partner_iso3c == "ISL") %>%
  mutate(importer_iso3c = ifelse(importer_iso3c == "ISL", "FRO", importer_iso3c),
         partner_iso3c = ifelse(partner_iso3c == "ISL", "FRO", partner_iso3c))

oils_meals_import_props <- rbind(oils_meals_import_props, iceland_faroe_props) %>%
    group_by(importer_iso3c, type) %>%
  mutate(total_import = sum(qty, na.rm = TRUE)) %>% 
  ungroup() %>%
  mutate(prop_import = qty/total_import) %>% 
  mutate(type = ifelse(type == "fish meal", "fish meal, cut offs", "fish oil, cut offs"))

```

### step 3

Multiply the import %s by the demand tonnes to get ingredient tonnes imported per country
    - EX: NOR requires 100 tonnes of FO from trimmings, they import 60% from the USA, then 60 tonnes of FO ingredients in NOR comes from the USA for salmon. 

```{r}
## step 3

demand_trade <- country_demand %>%
  left_join(oils_meals_import_props, by = c("iso3c" = "importer_iso3c", "ingredient" = "type")) %>% 
  mutate(tonnes_ingredient_demand_traded = prop_import*ingredient_demand)

sum(demand_trade$tonnes_ingredient_demand_traded, na.rm = TRUE)
# 1105093
sum(country_demand$ingredient_demand)
# 1105093

# perfect
```

### step 4

Group by exporter country to get total amount of FO or FM ingredients from trimmings they export under each diet demand scenario
    - EX: USA exports 60 tonnes to NOR, 40 tonnes to CHL, then they export 100 tonnes of FO from trimmings total for salmon feed. 
    
```{r}

## step 4
demand_trade_origin <- demand_trade %>%
  group_by(partner_iso3c, ingredient, diet, fcr_type) %>%
  summarise(total_ingredient_tonnes_traded = sum(tonnes_ingredient_demand_traded, na.rm = TRUE)) %>% 
  ungroup()

sum(demand_trade_origin$total_ingredient_tonnes_traded, na.rm = TRUE) # 1105093 good, the same as above. 
 
```

### Step 5 

Determine each countries' breakdown of trimmings species catch by flag country (e.g., where the catch is actually landed, rather than caught). Filter for all trimmings species (from biomar report), group by country and species, and determine the within-country proportions of catch from each species. For example, 55% of the trimmings species weight that USA lands is from XX spp, 24% from YY spp, etc...
 
```{r}
## step 5 

#get xyz file for the FAO areas in 0.5 resolution (catch data resolution)
fao_areas_shp <- 
  read_sf("https://www.fao.org/fishery/geoserver/fifao/ows?service=WFS&request=GetFeature&version=1.0.0&typeName=fifao:FAO_AREAS_CWP_NOCOASTLINE&outputFormat=json") |> filter(F_LEVEL == "MAJOR") |> 
  mutate(F_AREA = as.numeric(F_AREA)) |> 
  group_by(F_AREA) |> 
  group_split()

fao_xyz <- 
  
  map_df(.x = fao_areas_shp, .f = \(this_area){
  
  this_area_shp <- this_area
  
  this_area_raster <- terra::rasterize(x = vect(this_area_shp), y = rast(res = 0.5), field = this_area_shp$F_AREA)
  
  this_area_xyz <- terra::as.data.frame(x = this_area_raster, xy = TRUE)  |> mutate(fao_area_name = this_area_shp$NAME_EN)
  
  return(this_area_xyz)
  
}) |> rename(fao_area_code = layer,
             LonCentre = x,
             LatCentre = y)

# read in gear data and classify gears into benthic or pelagic
gear_df <- read.csv(file.path(watson_dir, "v5.0/Codes_gear.csv")) %>%
  janitor::clean_names()

gear_clean_df <- gear_df %>%
  dplyr::select(gear, f_gear_code, f_gear_label, fleet_gear_name) %>%
  distinct() %>%
   mutate(f_gear_label = tolower(f_gear_label),
         fleet_gear_name = tolower(fleet_gear_name),
         fleet_gear_name = str_replace(fleet_gear_name, 'tuna', ' tuna'),
         fleet_gear_name = str_replace_all(fleet_gear_name, '[^a-z]+', '_'))

benthic_gears <- c('trawl', 'dredge', 'trap')

trimmings_spp <- read_csv(here("data/raw_data/biomar/trimmings_spp_list.csv")) %>% 
  filter(!(common_name %in% c("Bali sardinella", "Goldstripe sardinella", "Japanese sardinella"))) # i supplemented this to include all sardine and anchovy species, as the biomar report just reports "Sardinella spp" and "Anchoveta" as a wide group. However, they specifically say that pacific sardines are not used for trimmings, so i will filter those species out here

watson_cells <- read.csv(file.path(watson_dir, "v5.0/Codes_cells.csv")) %>%
  left_join(fao_xyz)


codes_country <- read.csv(file.path(watson_dir, "v5.0/Codes_country.csv")) # read in country codes 


# calculate forage catch and add iso3c codes

trimmingsfish <- 
  watson_raw |> 
  mutate(trimmingsfish = ifelse(CommonName %in% c(trimmings_spp$common_name) | TaxonName %in% c(trimmings_spp$scientific_name), 1, NA)) %>%
  left_join(watson_cells)

trimmingsfish_cats <- trimmingsfish %>%
  filter(trimmingsfish == 1) %>%
  distinct(CommonName, TaxonName)

# write csv of trimmings fish species and add species specific FAO areas from BioMar sustainability report
write.csv(trimmingsfish_cats, here("prep/02_feed/data/trimmingsfish_species_watson.csv"), row.names = FALSE)
# now i've exported this to excel and manually added each FAO area and reading back in 
trimmingsfish_cats <- read.csv(here("prep/02_feed/data/trimmingsfish_species_fao_areas.csv"))

  trimmingsfish <- trimmingsfish %>% 
  mutate(trimmingsfish = 
           case_when(
             trimmingsfish == 1 & CommonName %in% c("Atlantic mackerel") & fao_area_code %in% c(27, 34) ~ 1,
             trimmingsfish == 1 & CommonName %in% c("European pilchard") & fao_area_code %in% c(27, 34, 37, 87) ~ 1,
           trimmingsfish == 1 & CommonName %in% c("Skipjack tuna", "Albacore", "Yellowfin tuna", "Tuna") & fao_area_code %in% c(87, 57) ~ 1,
           trimmingsfish == 1 & CommonName %in% c("European anchovy", "Sardinella", "Brazilian sardinella") & fao_area_code %in% c(34, 37) ~ 1,
           trimmingsfish == 1 & CommonName %in% c("Chub mackerel") & fao_area_code %in% c(87, 77, 71) ~ 1,
           trimmingsfish == 1 & CommonName %in% c("Anchovies", "Japanese anchovy", "Pacific anchoveta", "Indo-Pacific anchovies", "Californian anchovy", "Cape anchovy", "Atlantic anchoveta") & fao_area_code %in% c(47, 77, 37, 27, 34, 87, 61) ~ 1,
          trimmingsfish == 1 & CommonName %in% c("Atlantic herring", "Atlantic cod", "Capelin", "Blue whiting") & fao_area_code %in% c(27) ~ 1,
          trimmingsfish == 1 & CommonName %in% c("Alaska pollack") & fao_area_code %in% c(67) ~ 1,
          trimmingsfish == 1 & CommonName %in% c("Anchoveta") & fao_area_code %in% c(87) ~ 1,
          trimmingsfish == 1 & CommonName %in% c("Round sardinella", "Madeiran sardinella") & fao_area_code %in% c(27, 34, 37, 87) ~ 1,
          trimmingsfish == 1 & CommonName %in% c("Herrings sardines anchovies") & fao_area_code %in% c(34, 37, 47, 77, 27, 87, 61) ~ 1,
          TRUE ~ NA)) %>%
  left_join(watson_rgn, by = c("CountryName" = "country", "CNumber")) %>%
  mutate(catch = ReportedIND + ReportedNIND) %>%
  mutate(trimmings_catch = catch*trimmingsfish) %>%
  mutate(trimmings_catch = ifelse(is.na(trimmings_catch), 0, trimmings_catch)) %>%
  left_join(gear_clean_df, by = c("Gear" = "gear", "FGearCode"= "f_gear_code")) %>%
    mutate(water_col_position = ifelse(fleet_gear_name %in% benthic_gears, "benthic", "pelagic")) %>%
  mutate(prop_discards = (DiscardsIND + DiscardsNIND)/ (DiscardsIND + DiscardsNIND + trimmings_catch)) %>%
  mutate(prop_discards = ifelse(trimmings_catch != 0, prop_discards, 0))



write_rds(trimmingsfish, file.path(watson_dir, "v5.0/int/all_catch_trimmingsfish.rds"))

# trimmingsfish <- readRDS( file.path(watson_dir, "v5.0/int/all_catch_trimmingsfish.rds"))

## get an estimate of proportion of forage fish catch coming from different species in different landing countries.
trimmings_catch_iso <- trimmingsfish %>%
  filter(trimmings_catch > 0) %>%
  group_by(iso3c, CommonName, TaxonName) %>%
  summarise(catch = sum(trimmings_catch, na.rm = TRUE)) %>%
  ungroup() %>%
  group_by(iso3c) %>%
  mutate(total_fofm_trim_catch = sum(catch, na.rm = TRUE)) %>%
  ungroup() %>%
  mutate(prop_of_iso_total = catch/total_fofm_trim_catch)

unique(trimmings_catch_iso$CommonName) # much better, 

setdiff(trimmings_spp$scientific_name, unique(trimmings_catch_iso$TaxonName)) # [1] "Thunnini sp"         "Sardinella spp"      "Engraulis spp"       ".. many of the common names associated with these match though.

unique(trimmings_catch_iso$TaxonName)


```


### step 6 

Apply these within country percentages to total country exported fish oil and fish meal (calculated in step 4). Now we will have X tonnes of fish oil and Y tonnes of fish meal exported from or kept within country Z per trimmings fish species per diet scenario. 
     - For example, if 55% of the trimmings catch in USA is from XX spp, then 55% of the fish meal (and 55% of the fish oil) exported from or kept in USA will be attributed to XX spp. 
     
```{r}
## step 6 
trimmings_fofm_origins <- trimmings_catch_iso %>%
  dplyr::select(iso3c, CommonName, TaxonName, prop_of_iso_total) %>%
  left_join(demand_trade_origin, by = c("iso3c" = "partner_iso3c")) %>%
  group_by(iso3c, ingredient, diet, fcr_type) %>%
  mutate(total_ingredient = sum(total_ingredient_tonnes_traded)) %>%
  ungroup() %>%
  mutate(tonnes_fofm_trimmings_spp = total_ingredient_tonnes_traded*prop_of_iso_total) %>%
  dplyr::select(iso3c, CommonName, TaxonName, ingredient, diet, fcr_type, tonnes_fofm_trimmings_spp) %>%
  filter(!is.na(tonnes_fofm_trimmings_spp))

sum(trimmings_fofm_origins$tonnes_fofm_trimmings_spp, na.rm = TRUE) # 1087952 losing ~13k bc of Uruguay is missing... they don't report catch for any of the trimmings spp we have... 


trimmings_test <- trimmings_fofm_origins %>%
  group_by(iso3c, ingredient, diet, fcr_type) %>%
  summarise(total = sum(tonnes_fofm_trimmings_spp))

setdiff(demand_trade_origin$partner_iso3c, trimmings_test$iso3c) # [1] "BRN" "COD" "CRI" "HND" "KWT" "MAC" "SAU" "SOM" "URY"

trim_test_2 <- demand_trade_origin %>% 
  filter(partner_iso3c %in% c(setdiff(demand_trade_origin$partner_iso3c, trimmings_test$iso3c))) # yup this is because of these missing countries, without trimming spp reported in catch data...

```


### step 7 

Next, we can use species specific allocation rates to convert the fish oil and fish meal from trimmings to raw material live weight equivalents. Now we have total tonnes of fish oil and fish meal live weight equivalents per landing country per species. 
    - for adjustments: 
       - If possible use species rates
       - Otherwise, use a global average of all species 

```{r}

## step 7

## read in allocation data 
trimmings_allocation_raw <- read_xlsx(here("data/tidy_data/allocation/trimmings_allocation_factors.xlsx"), sheet = 3) %>%
  dplyr::select(CommonName = common_name, TaxonName = sci_name, ingredient, mass_byproduct_cf, ge_byproduct_cf, econ_byproduct_cf)

## missing tuna... take means of tuna spp: Skipjack tuna, Yellowfin tuna

trimmings_allocation_tidy <- trimmings_allocation_raw %>% 
  mutate(ingredient = ifelse(ingredient == "Fishmeal", "fish meal, cut offs", "fish oil, cut offs")) %>%
  dplyr::select(CommonName, TaxonName, ingredient, ge_value = ge_byproduct_cf, mass_value = mass_byproduct_cf, econ_value = econ_byproduct_cf) 

tuna_mean <- trimmings_allocation_tidy %>%
  filter(CommonName %in% c("Skipjack tuna", "Yellowfin tuna")) %>%
  mutate(CommonName = "Tuna",
         TaxonName = "Thunnus") %>%
  group_by(CommonName, TaxonName, ingredient) %>%
  summarise(ge_value = mean(ge_value, na.rm = TRUE),
            mass_value = mean(mass_value, na.rm = TRUE),
            econ_value = mean(econ_value, na.rm = TRUE)) %>%
  ungroup()

trimmings_allocation_tidy <- rbind(trimmings_allocation_tidy, tuna_mean) 


## will need to gapfill all anchovies by anchovy mean 
## gapfill sardine by sardine mean 

global_averages <- trimmings_allocation_tidy %>%
  group_by(ingredient) %>%
  summarise(average_ge = mean(ge_value, na.rm = TRUE),
            average_mass = mean(mass_value, na.rm = TRUE),
            average_econ = mean(econ_value, na.rm = TRUE)) %>%
  ungroup()

ga_mass_FO <- global_averages %>%
  filter(ingredient == "fish oil, cut offs") %>%
  dplyr::pull(average_mass)

ga_mass_FM <- global_averages %>%
  filter(ingredient == "fish meal, cut offs") %>%
    dplyr::pull(average_mass)

  
ga_ge_FO <- global_averages %>%
  filter(ingredient == "fish oil, cut offs") %>%
    dplyr::pull(average_ge)


ga_ge_FM <- global_averages %>%
  filter(ingredient == "fish meal, cut offs") %>%
  dplyr::pull(average_ge)

ga_econ_FO <- global_averages %>%
  filter(ingredient == "fish oil, cut offs") %>%
    dplyr::pull(average_econ)


ga_econ_FM <- global_averages %>%
  filter(ingredient == "fish meal, cut offs") %>%
  dplyr::pull(average_econ)


trimmings_allocations_df <- trimmings_allocation_tidy %>%
  group_by(CommonName, TaxonName, ingredient) %>% # lets just do global averages per species 
  summarise(ge_value = mean(ge_value, na.rm = TRUE),
            mass_value = mean(mass_value, na.rm = TRUE),
            econ_value = mean(econ_value, na.rm = TRUE)) %>%
  ungroup() %>%
  filter(!is.na(ge_value))  %>%
  add_row(CommonName = "Chub mackerel", TaxonName = "Scomber japonicus", ingredient = "fish meal, cut offs", ge_value = 1.037364765, mass_value = 1.122715405, econ_value = 0.10313204) %>%
    add_row(CommonName = "Chub mackerel", TaxonName = "Scomber japonicus", ingredient = "fish oil, cut offs", ge_value = 2.038421764, mass_value = 1.5021600, econ_value = 0.134279301) %>%
  mutate(TaxonName = ifelse(TaxonName == "Gadus chalcogrammus", "Theragra chalcogramma", TaxonName)) %>%
  mutate(CommonName = ifelse(TaxonName == "Engraulis ringens", "Anchoveta", CommonName))

# this is annoying.. but we need to gapfill the anchovies and sardine species
anchovy_mean_fm_ge <- trimmings_allocations_df %>% 
  filter(CommonName == "Anchoveta", ingredient == "fish meal, cut offs") %>% 
  pull(ge_value)

anchovy_mean_fo_ge <- trimmings_allocations_df %>% 
  filter(CommonName == "Anchoveta", ingredient == "fish oil, cut offs") %>% 
  pull(ge_value)

anchovy_mean_fm_mass <- trimmings_allocations_df %>% 
  filter(CommonName == "Anchoveta", ingredient == "fish meal, cut offs") %>% 
  pull(mass_value)

anchovy_mean_fo_mass <- trimmings_allocations_df %>% 
  filter(CommonName == "Anchoveta", ingredient == "fish oil, cut offs") %>% 
  pull(mass_value)

anchovy_mean_fm_econ <- trimmings_allocations_df %>% 
  filter(CommonName == "Anchoveta", ingredient == "fish meal, cut offs") %>% 
  pull(econ_value)

anchovy_mean_fo_econ <- trimmings_allocations_df %>% 
  filter(CommonName == "Anchoveta", ingredient == "fish oil, cut offs") %>% 
  pull(econ_value)


sardine_mean_fm_ge <- trimmings_allocations_df %>%
  filter(CommonName == "Atlantic sardine", ingredient == "fish meal, cut offs") %>%
  pull(ge_value)

sardine_mean_fo_ge <- trimmings_allocations_df %>%
  filter(CommonName == "Atlantic sardine", ingredient == "fish oil, cut offs") %>%
  pull(ge_value)

sardine_mean_fm_mass <- trimmings_allocations_df %>%
  filter(CommonName == "Atlantic sardine", ingredient == "fish meal, cut offs") %>%
  pull(mass_value)

sardine_mean_fo_mass <- trimmings_allocations_df %>%
  filter(CommonName == "Atlantic sardine", ingredient == "fish oil, cut offs") %>%
  pull(mass_value)

sardine_mean_fm_econ <- trimmings_allocations_df %>%
  filter(CommonName == "Atlantic sardine", ingredient == "fish meal, cut offs") %>%
  pull(econ_value)

sardine_mean_fo_econ <- trimmings_allocations_df %>%
  filter(CommonName == "Atlantic sardine", ingredient == "fish oil, cut offs") %>%
  pull(econ_value)



## Match allocation data to the fish oil and fish meal per species from comtrade

trimmings_fofm_allocation <- trimmings_fofm_origins %>%
  left_join(trimmings_allocations_df, by = c("ingredient", "CommonName", "TaxonName")) %>%
  mutate(ge_value = case_when(
    is.na(ge_value) & str_detect(CommonName, "anchovy|Anchovies|anchovies|") & ingredient == "fish meal, cut offs" ~ anchovy_mean_fm_ge,
    is.na(ge_value) & str_detect(CommonName, "anchovy|Anchovies|anchovies|") & ingredient == "fish oil, cut offs" ~ anchovy_mean_fo_ge,
  
      is.na(ge_value) & str_detect(CommonName, "Sard|sard") & ingredient == "fish meal, cut offs" ~ sardine_mean_fm_ge,
    is.na(ge_value) & str_detect(CommonName, "Sard|sard") & ingredient == "fish oil, cut offs" ~ sardine_mean_fo_ge,
  TRUE ~ ge_value
  ),
  mass_value = case_when(
      is.na(mass_value) & str_detect(CommonName, "anchovy|Anchovies|anchovies|") & ingredient == "fish meal, cut offs" ~ anchovy_mean_fm_mass,
   is.na(mass_value) & str_detect(CommonName, "anchovy|Anchovies|anchovies|") & ingredient == "fish oil, cut offs" ~ anchovy_mean_fo_mass, 
     is.na(mass_value) & str_detect(CommonName, "Sard|sard") & ingredient == "fish meal, cut offs" ~ sardine_mean_fm_mass,
   is.na(mass_value) & str_detect(CommonName, "Sard|sard") & ingredient == "fish oil, cut offs" ~ sardine_mean_fo_mass,
   TRUE ~ mass_value
  ),
   econ_value = case_when(
      is.na(econ_value) & str_detect(CommonName, "anchovy|Anchovies|anchovies|") & ingredient == "fish meal, cut offs" ~ anchovy_mean_fm_econ,
   is.na(econ_value) & str_detect(CommonName, "anchovy|Anchovies|anchovies|") & ingredient == "fish oil, cut offs" ~ anchovy_mean_fo_econ, 
     is.na(econ_value) & str_detect(CommonName, "Sard|sard") & ingredient == "fish meal, cut offs" ~ sardine_mean_fm_econ,
   is.na(econ_value) & str_detect(CommonName, "Sard|sard") & ingredient == "fish oil, cut offs" ~ sardine_mean_fo_econ,
   TRUE ~ econ_value
  )) %>%
  mutate(ge_value = case_when(
    is.na(ge_value) & ingredient == "fish meal, cut offs" ~ ga_ge_FM,
    is.na(ge_value) & ingredient == "fish oil, cut offs" ~ ga_ge_FO, 
    TRUE ~ ge_value),
    mass_value = case_when(
      is.na(mass_value) & ingredient == "fish meal, cut offs" ~ ga_mass_FM,
      is.na(mass_value) & ingredient == "fish oil, cut offs" ~ ga_mass_FO, 
      TRUE ~ mass_value
    ),
    econ_value = case_when(
      is.na(econ_value) & ingredient == "fish meal, cut offs" ~ ga_econ_FM,
      is.na(econ_value) & ingredient == "fish oil, cut offs" ~ ga_econ_FO, 
      TRUE ~ econ_value
    )
  ) %>%
  pivot_longer(cols = c("ge_value", "mass_value", "econ_value"), 
               names_to = "allocation_type", 
               values_to = "allocation_value"
               ) %>%
  mutate(allocation_type = case_when(
    allocation_type == "ge_value" ~ "ge",
    allocation_type == "mass_value" ~ "mass",
    TRUE ~ "economic"
  )) %>%
  mutate(tonnes_fofm_trimmings_spp_live_weight = tonnes_fofm_trimmings_spp*allocation_value) %>%
  dplyr::select(iso3c, CommonName, TaxonName, ingredient, diet, fcr_type, tonnes_fofm_trimmings_spp, allocation_type, allocation_value, tonnes_fofm_trimmings_spp_live_weight)

sum(trimmings_fofm_allocation$tonnes_fofm_trimmings_spp_live_weight) # 5100000/3 = 1700000 Ok this makes sense. Weird that it is an exactly even number.. but the numbers check out... see below. 

sum(trimmings_fofm_allocation$tonnes_fofm_trimmings_spp) # 3263857/3 = 1087952 Ok this makes sense. 

mean(trimmings_fofm_allocation$allocation_value) # 1.459467

1700000/1087952 # 1.562569 seems fine 

```


### step 8

Divide the tonnes of fish oil and fish meal live weight equivalents by the total amount of trimmings catch in each country. This gives us a proportion of trimmings fish catch in each country that goes to FMFO. 

```{r}

## step 8

# Divide the tonnes of fish oil and fish meal live weight equivalents by the total amount of trimmings catch in each country. This gives us a proportion of trimmings fish catch in each country that goes to FMFO. 

trimmings_rgn_totals <- trimmings_catch_iso %>%
  group_by(iso3c, TaxonName) %>%
  summarise(trim_fofm_catch = sum(catch)) %>%
  ungroup()


trimmings_fofm_allocation_totals <- trimmings_fofm_allocation %>%
  group_by(iso3c, ingredient, TaxonName, diet, fcr_type, allocation_type) %>%
  summarise(total_tonnes_fofm_lw = sum(tonnes_fofm_trimmings_spp_live_weight, na.rm = TRUE)) %>%
              ungroup()

trim_props <- left_join(trimmings_rgn_totals, trimmings_fofm_allocation_totals) %>%
  filter(!is.na(diet)) %>% # filter out NAs since these countries dont produce any trimmings catch for salmon feeds
  mutate(prop = ifelse(trim_fofm_catch == 0, 0, total_tonnes_fofm_lw/trim_fofm_catch)) %>%
  dplyr::select(iso3c, ingredient, TaxonName, diet, fcr_type, allocation_type, prop) %>%
  filter(!is.na(iso3c)) # weird iso3c with all 0s 

write_rds(trim_props, here("prep/02_feed/data/embodied_FMFO_trimmings_rgns.rds"))


trim_props <- readRDS(here("prep/02_feed/data/embodied_FMFO_trimmings_rgns.rds"))
trimmingsfish <- readRDS(file.path(watson_dir, "v5.0/int/all_catch_trimmingsfish.rds"))
watson_cells <- read.csv(file.path(watson_dir, "v5.0/Codes_cells.csv"))

diets <- unique(trim_props$diet)
allocations <- unique(trim_props$allocation_type)
ingredients <- unique(trim_props$ingredient)
depth_positions <- c("benthic", "pelagic")
# fcrs <- unique(trim_props$fcr_type) # we are only running regular FCRs now

for(diet_type in diets){
  for(fcr in c("regular")){
  for(allocation in allocations){
    for(ingredient_type in ingredients){
      for(pos in depth_positions){

    # diet_type = "fish-dominant"
    # allocation = "economic"
    # ingredient_type = "fish meal, cut offs"
    #  pos = "pelagic"
    #   fcr = "regular"
        
        ## comment this out if u want to skip those already run
      if(all(file.exists(sprintf(here("prep/02_feed/output/%s/%s/trimmings fish_%s_%s_%s_catch_P.tif"), diet_type, fcr, ingredient_type, allocation, pos), sprintf(here("prep/02_feed/output/%s/%s/trimmings fish_%s_%s_%s_bycatch_P.tif"), diet_type, fcr, ingredient_type, allocation, pos)))){

          print(cat("rasters exists, skipping!"))

          next()
        }

        
  watson_raster_template <- terra::rast(ncol=720, nrow=360, vals=c(1:259200))
  
  system_prop <- trim_props %>%
    dplyr::filter(diet == diet_type,
                  allocation_type == allocation,
                  ingredient == ingredient_type, 
                  fcr_type == fcr) %>%
    dplyr::select(iso3c, diet, fcr_type, TaxonName, ingredient, prop)
  
  trimmingsfish_system_catch <- left_join(trimmingsfish, system_prop) %>%
    filter(!is.na(ingredient)) %>%
    mutate(prop = ifelse(is.na(prop), 0, prop)) %>%
    mutate(trimmings_catch_foodsystem = prop*trimmings_catch) %>%
    filter(water_col_position == pos) %>% ## filter for DEPTH RANGE HERE
    group_by(Cell) %>%
    summarise(tonnes = sum(trimmings_catch_foodsystem, na.rm = TRUE)) %>%
    ungroup() %>%
    left_join(watson_cells) %>%
    dplyr::select(LonCentre, LatCentre, tonnes) %>%
    filter(tonnes > 0)
  

  trimmings_fish_system_catch <- rast(trimmingsfish_system_catch, type = "xyz", crs = crs(watson_raster_template)) %>%
    project(watson_raster_template)

# plot(log(trimmings_fish_system_catch + 1))

    writeRaster(trimmings_fish_system_catch, sprintf(here("prep/02_feed/output/%s/%s/trimmings fish_%s_%s_%s_catch_P.tif"), diet_type, fcr, ingredient_type, allocation, pos), overwrite = TRUE)
    
    
      trimmingsfish_system_bycatch <- left_join(trimmingsfish, system_prop) %>%
    filter(!is.na(ingredient)) %>%
    mutate(prop = ifelse(is.na(prop), 0, prop)) %>%
    mutate(trimmings_catch_foodsystem = prop*trimmings_catch*prop_discards) %>%
    filter(water_col_position == pos) %>% ## filter for DEPTH RANGE HERE
    group_by(Cell) %>%
    summarise(tonnes = sum(trimmings_catch_foodsystem, na.rm = TRUE)) %>%
    ungroup() %>%
    left_join(watson_cells) %>%
    dplyr::select(LonCentre, LatCentre, tonnes)
  

  trimmings_fish_system_bycatch <- rast(trimmingsfish_system_bycatch, type = "xyz", crs = crs(watson_raster_template)) %>%
    project(watson_raster_template)

# plot(log(trimmings_fish_system_bycatch + 1))

    writeRaster(trimmings_fish_system_bycatch, sprintf(here("prep/02_feed/output/%s/%s/trimmings fish_%s_%s_%s_bycatch_P.tif"), diet_type, fcr, ingredient_type, allocation, pos), overwrite = TRUE)
    
    print(global(trimmings_fish_system_bycatch, "sum", na.rm = TRUE)/global(trimmings_fish_system_catch, "sum", na.rm = TRUE))
    
    
  
      }
    }
   }
  }
}

```

## Data check 

```{r}
trim_files <- list.files(here("prep/02_feed/output/fish-dominant/regular"), pattern = "fish oil|fish meal|trimmings", full.names = TRUE)

trim_files <- trim_files[grep("_P", trim_files)]
trim_files <- trim_files[grep("trimmings fish", trim_files)]
trim_files <- trim_files[grep("economic", trim_files)]

new_rasts <- rast(trim_files)

new_names <- basename(trim_files)

names(new_rasts) <- new_names

# plot(log(new_rasts+1))

global(new_rasts, "sum", na.rm = TRUE)

test <- app(new_rasts, "sum", na.rm = TRUE)

test[test == 0] <- NA


test <- trimmings_fofm_allocation %>%
  group_by(diet,
         allocation_type,
         ingredient, fcr_type) %>%
  summarise(sum = sum(tonnes_fofm_trimmings_spp_live_weight)) %>%
  ungroup()
test

```
