==> prep/02_feed/6_resample_feed_rasts.Rmd <==
---
title: "Resample all crop and fish based feed rasters to 100km2 cells"
output: html_document
editor_options: 
  chunk_output_type: console
---

## Summary 

Goal: Resample all crop and fish based feed rasters to 100km2 cells

Note: I am saving all of these rasters on our RDSI server. I was encountering problems with trying to save so many files on github.


```{r}

library(here)
library(tidyverse)
library(countrycode)
library(tidytext)
library(knitr)
library(strex)
library(sjmisc)

source(here("src/fxns.R"))
source(here("src/directories.R"))
source(here("src/spatial.R"))

moll_crs <- crs(moll_template)

biodiv_dir <- file.path(rdsi_dir, "biodiversity_impacts")

```

## Methods 

We will present data at at ~10km resolution in the WGS84 Mollweide coordinate reference system, an equal area projection that retains considerable accuracy in area at polar latitudes.

Reproject and aggregate the ingredient crop rasters. It runs quick since they are already ~9km by 9km.

```{r}

for(diet in c("fish-dominant", "plant-dominant")){
  for(fcr in c("regular")){
  
  # diet = "fish-dominant"
  #   fcr = "regular"
directory <- sprintf("prep/02_feed/output/%s/%s", diet, fcr)

files <- list.files(here(directory), full.names = TRUE)

filtered_files <- files[!grepl("fish oil|fish meal|archive", files)]

for(i in 1:length(filtered_files)){
  # i = 17
  file <- filtered_files[i]
  file_name <- basename(file)
  
  crop_rast <- rast(file)
  
  area_rast <- cellSize(crop_rast, unit = "km")
  
  crop_rast_fix <- crop_rast/area_rast
  
  crop_reproject <- project(crop_rast_fix, moll_template, method = "bilinear")
  
  crop_reproject_area <- cellSize(crop_reproject, unit = "km")
  
  crop_reproject_fix <- crop_reproject*crop_reproject_area
  
  print(global(crop_reproject_fix, "sum", na.rm = TRUE)) # layer 20374.35
  print(global(crop_rast, "sum", na.rm = TRUE)) # layer 20373.72
  
  ## perfect... 
  writeRaster(crop_reproject_fix, sprintf(file.path(biodiv_dir, "int/resampled_ingredient_rasts//%s/%s/%s"), diet,fcr, file_name), overwrite = TRUE)
  }

  }
}

```


Reproject and disaggregate the ingredient fish rasters to 10km res.

```{r}

for(diet in c("fish-dominant", "plant-dominant")){
  for(fcr in c("regular")){
  
  # diet = "fish-dominant"
  #  fcr = "regular"
directory <- sprintf("prep/02_feed/output/%s/%s", diet, fcr)

files <- list.files(here(directory), full.names = TRUE)

filtered_files <- files[grepl("fish oil|fish meal", files)]
filtered_files <- filtered_files[grepl("trimmings|forage fish", filtered_files)]
# filtered_files <- filtered_files[grepl("economic", filtered_files)]
# filtered_files <- filtered_files[grepl("_A", filtered_files)]

# filtered_files <- filtered_files[grep("_P", filtered_files)]

# 2*2*3*2*2 = 48 files.. makes sense

for(i in 1:length(filtered_files)){
  #i = 8
  file <- filtered_files[i]
  file_name <- basename(file)
  
  fish_rast <- rast(file)
  
  area_rast <- cellSize(fish_rast, unit = "km")
  
  fish_rast_fix <- fish_rast/area_rast
  
  fish_reproject <- project(fish_rast_fix, moll_template, method = "bilinear")
  
  fish_reproject_area <- cellSize(fish_reproject, unit = "km")
  
  fish_reproject_fix <- fish_reproject*fish_reproject_area
  
  print(global(fish_reproject_fix, "sum", na.rm = TRUE)) # layer 3375102
  print(global(fish_rast, "sum", na.rm = TRUE)) # layer 3381631
  
  ## perfect... 
  writeRaster(fish_reproject_fix, sprintf(file.path(biodiv_dir, "int/resampled_ingredient_rasts/%s/%s/%s"), diet, fcr, file_name), overwrite = TRUE)
  
    }
  }
}


test_fis <- list.files(here("prep/02_feed/output/fish-dominant"), full.names = TRUE, recursive = TRUE)
test_fis_res <-  list.files(file.path(biodiv_dir, "int/resampled_ingredient_rasts/fish-dominant"), full.names = TRUE, recursive = TRUE)

## check 

test_plant <- list.files(here("prep/02_feed/output/plant-dominant"), full.names = TRUE, recursive = TRUE)
test_plant_res <-  list.files(file.path(biodiv_dir, "int/resampled_ingredient_rasts/plant-dominant"), full.names = TRUE, recursive = TRUE)

# perfect 



test_reg <- rast(file.path(biodiv_dir, "int/resampled_ingredient_rasts/fish-dominant/regular/forage fish_fish meal_economic_pelagic_catch_A.tif"))
  
test_eff <- rast(file.path(biodiv_dir, "int/resampled_ingredient_rasts/fish-dominant/efficient/forage fish_fish meal_economic_pelagic_catch_A.tif"))

diff <- (test_reg - test_eff)/test_reg

plot(diff) # ok cool

```


Data check 

```{r}
# take a look at plant dominant

test_names <- str_replace_all(str_before_last(list.files(file.path(biodiv_dir, "int/resampled_ingredient_rasts/plant-dominant/regular"), pattern = "economic_A.tif"), ".tif"), "_", ":")

test <- rast(list.files(file.path(biodiv_dir, "int/resampled_ingredient_rasts/plant-dominant/regular"), pattern = "economic_A.tif", full.names = TRUE))

names(test) <- test_names

# plot(log(test+1))

test_df <- as.data.frame(global(test, "sum", na.rm = TRUE))

test_df %>% arrange(-sum)

#                                                   sum
# Rapeseed:canola oil:economic:A             3350.17145
# Soybean:soy protein concentrate:economic:A 2482.72642
# Wheat:wheat gluten:economic:A              1954.18908
# Pulses:pea protein concentrate:economic:A   572.73846
# Pulses:faba beans:economic:A                570.68376
# Wheat:wheat:economic:A                      507.38673
# Pulses:pea flour:economic:A                 466.24491
# CropsNES:linseed oil:economic:A             439.87800
# Sunflower:sunflower meal:economic:A         315.45644
# Pulses:guar meal:economic:A                 305.56726
# CropsNES:coconut oil:economic:A              81.27246
# Soybean:soy oil:economic:A                   73.95469
# Maize:corn gluten meal:economic:A            14.18603

# take a look at fish dominant
test_names <- str_replace_all(str_before_last(list.files(file.path(biodiv_dir, "int/resampled_ingredient_rasts/fish-dominant/regular"), pattern = "economic_A.tif"), ".tif"), "_", ":")

test <- rast(list.files(file.path(biodiv_dir, "int/resampled_ingredient_rasts/fish-dominant/regular"), pattern = "economic_A.tif", full.names = TRUE))

names(test) <- test_names

# plot(log(test+1))

test_df <- as.data.frame(global(test, "sum", na.rm = TRUE))

test_df %>% arrange(-sum)
#                                          sum
# Wheat:wheat gluten:economic:A     2592.29164
# Soybean:soybean meal:economic:A   1578.01008
# Wheat:wheat:economic:A             546.41648
# Pulses:faba beans:economic:A       317.04654
# Maize:corn gluten meal:economic:A   81.06304

```
